@@ -1,82 +1,109 @@
 package db
 
 import (
+	"database/sql"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"os"
-	"path"
-	"strings"
+	"path/filepath"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/rqlite/rqlite/command"
-	"github.com/rqlite/rqlite/command/encoding"
-	"github.com/rqlite/rqlite/db/testdata"
-	"github.com/rqlite/rqlite/testdata/chinook"
+	"github.com/rqlite/rqlite/v8/command/encoding"
+	command "github.com/rqlite/rqlite/v8/command/proto"
 )
 
-/*
- * Lowest-layer database tests
- */
-
-func Test_DbFileCreation(t *testing.T) {
-	dir := t.TempDir()
-	dbPath := path.Join(dir, "test_db")
-
-	db, err := Open(dbPath, false)
+// Test_OpenNonExistentDatabase tests that opening a non-existent database
+// works OK. It should.
+func Test_OpenNonExistentDatabase(t *testing.T) {
+	path := mustTempPath()
+	defer os.Remove(path)
+	_, err := Open(path, false, false)
 	if err != nil {
-		t.Fatalf("failed to open new database: %s", err.Error())
+		t.Fatalf("error opening non-existent database")
 	}
-	if db == nil {
-		t.Fatal("database is nil")
-	}
-	if db.InMemory() {
-		t.Fatal("on-disk database marked as in-memory")
-	}
-	if db.FKEnabled() {
-		t.Fatal("FK constraints marked as enabled")
-	}
-	if db.Path() != dbPath {
-		t.Fatal("database path is incorrect")
-	}
-
-	if _, err := os.Stat(dbPath); os.IsNotExist(err) {
-		t.Fatalf("%s does not exist after open", dbPath)
-	}
-	err = db.Close()
-	if err != nil {
-		t.Fatalf("failed to close database: %s", err.Error())
+	// Confirm a file was created.
+	if !fileExists(path) {
+		t.Fatalf("database file not created")
 	}
 }
 
-func Test_CompileOptions(t *testing.T) {
-	db, path := mustCreateDatabase()
+func Test_WALRemovedOnClose(t *testing.T) {
+	path := mustTempPath()
+	defer os.Remove(path)
+	db, err := Open(path, false, true)
+	if err != nil {
+		t.Fatalf("error opening non-existent database")
+	}
+	defer db.Close()
+	if !db.WALEnabled() {
+		t.Fatalf("WAL mode not enabled")
+	}
+
+	_, err = db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
+	if err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	walPath := db.WALPath()
+	if !fileExists(walPath) {
+		t.Fatalf("WAL file does not exist after creating a table")
+	}
+	if err := db.Close(); err != nil {
+		t.Fatalf("error closing database: %s", err.Error())
+	}
+
+	if fileExists(db.WALPath()) {
+		t.Fatalf("WAL file not removed after closing the database")
+	}
+}
+
+func Test_RemoveFiles(t *testing.T) {
+	d := t.TempDir()
+	mustCreateClosedFile(fmt.Sprintf("%s/foo", d))
+	mustCreateClosedFile(fmt.Sprintf("%s/foo-wal", d))
+
+	if err := RemoveFiles(fmt.Sprintf("%s/foo", d)); err != nil {
+		t.Fatalf("failed to remove files: %s", err.Error())
+	}
+
+	files, err := os.ReadDir(d)
+	if err != nil {
+		t.Fatalf("failed to read directory: %s", err.Error())
+	}
+	if len(files) != 0 {
+		t.Fatalf("expected directory to be empty, but wasn't")
+	}
+}
+
+func Test_DBPaths(t *testing.T) {
+	dbWAL, pathWAL := mustCreateOnDiskDatabaseWAL()
+	defer dbWAL.Close()
+	defer os.Remove(pathWAL)
+	if exp, got := pathWAL, dbWAL.Path(); exp != got {
+		t.Fatalf("expected path %s, got %s", exp, got)
+	}
+	if exp, got := pathWAL+"-wal", dbWAL.WALPath(); exp != got {
+		t.Fatalf("expected WAL path %s, got %s", exp, got)
+	}
+	if p1, p2 := WALPath(pathWAL), dbWAL.WALPath(); p1 != p2 {
+		t.Fatalf("WAL paths are not equal (%s != %s)", p1, p2)
+	}
+
+	db, path := mustCreateOnDiskDatabase()
 	defer db.Close()
 	defer os.Remove(path)
-
-	_, err := db.CompileOptions()
-	if err != nil {
-		t.Fatalf("failed to retrieve compilation options: %s", err.Error())
+	if exp, got := path, db.Path(); exp != got {
+		t.Fatalf("expected path %s, got %s", exp, got)
+	}
+	if exp, got := "", db.WALPath(); exp != got {
+		t.Fatalf("expected WAL path %s, got %s", exp, got)
 	}
 }
 
-func Test_TableNotExist(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	q, err := db.QueryStringStmt("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `[{"error":"no such table: foo"}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
+// Test_TableCreation tests basic operation of an database
 func Test_TableCreation(t *testing.T) {
-	db, path := mustCreateDatabase()
+	db, path := mustCreateOnDiskDatabaseWAL()
 	defer db.Close()
 	defer os.Remove(path)
 
@@ -88,52 +115,45 @@ func Test_TableCreation(t *testing.T) {
 		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
 	}
 
-	q, err := db.QueryStringStmt("SELECT * FROM foo")
+	testQ := func() {
+		t.Helper()
+		q, err := db.QueryStringStmt("SELECT * FROM foo")
+		if err != nil {
+			t.Fatalf("failed to query empty table: %s", err.Error())
+		}
+		if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]`, asJSON(q); exp != got {
+			t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+		}
+	}
+
+	_, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
 	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
+		t.Fatalf("error executing insertion into table: %s", err.Error())
 	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"]}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+	testQ()
+
+	// Confirm checkpoint works without error.
+	if err := db.Checkpoint(); err != nil {
+		t.Fatalf("failed to checkpoint database: %s", err.Error())
 	}
+	testQ()
+
+	// Check that VACUUM returns without error.
+	if err := db.Vacuum(); err != nil {
+		t.Fatalf("failed to VACUUM database: %s", err.Error())
+	}
+	testQ()
 }
 
-// Test_TableCreationInMemory tests basic operation of an in-memory database,
-// ensuring that using different connection objects (as the Execute and Query
-// will do) works properly i.e. that the connections object work on the same
-// in-memory database.
-func Test_TableCreationInMemory(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	if !db.InMemory() {
-		t.Fatal("in-memory database marked as not in-memory")
-	}
-
-	r, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-
-	q, err := db.QueryStringStmt("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"]}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-// Test_TableCreationInMemoryFK ensures foreign key constraints work
-func Test_TableCreationInMemoryFK(t *testing.T) {
+// Test_TableCreationFK ensures foreign key constraints work
+func Test_TableCreationFK(t *testing.T) {
 	createTableFoo := "CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)"
 	createTableBar := "CREATE TABLE bar (fooid INTEGER NOT NULL PRIMARY KEY, FOREIGN KEY(fooid) REFERENCES foo(id))"
 	insertIntoBar := "INSERT INTO bar(fooid) VALUES(1)"
 
-	db := mustCreateInMemoryDatabase()
+	db, path := mustCreateOnDiskDatabaseWAL()
 	defer db.Close()
+	defer os.Remove(path)
 
 	r, err := db.ExecuteStringStmt(createTableFoo)
 	if err != nil {
@@ -160,8 +180,10 @@ func Test_TableCreationInMemoryFK(t *testing.T) {
 	}
 
 	// Now, do same testing with FK constraints enabled.
-	dbFK := mustCreateInMemoryDatabaseFK()
+	dbFK, path := mustCreateOnDiskDatabaseWALFK()
 	defer dbFK.Close()
+	defer os.Remove(path)
+
 	if !dbFK.FKEnabled() {
 		t.Fatal("FK constraints not marked as enabled")
 	}
@@ -191,612 +213,10 @@ func Test_TableCreationInMemoryFK(t *testing.T) {
 	}
 }
 
-func Test_SQLiteMasterTable(t *testing.T) {
-	db, path := mustCreateDatabase()
+func Test_ConcurrentQueries(t *testing.T) {
+	db, path := mustCreateOnDiskDatabaseWAL()
 	defer db.Close()
 	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt("SELECT * FROM sqlite_master")
-	if err != nil {
-		t.Fatalf("failed to query master table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["type","name","tbl_name","rootpage","sql"],"types":["text","text","text","int","text"],"values":[["table","foo","foo",2,"CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-func Test_SQLiteTimeTypes(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo(d DATE, ts TIMESTAMP, dt DATETIME)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO foo(d, ts, dt) VALUES("2004-02-14", "2004-02-14 07:15:00", "2021-10-22 19:39:32.016087")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt("SELECT d, ts, dt FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query master table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["d","ts","dt"],"types":["date","timestamp","datetime"],"values":[["2004-02-14T00:00:00Z","2004-02-14T07:15:00Z","2021-10-22T19:39:32.016087Z"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-func Test_NotNULLField(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	r, err := db.QueryStringStmt(`PRAGMA table_info("foo")`)
-	if err != nil {
-		t.Fatalf("failed to get PRAGMA table_info: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["cid","name","type","notnull","dflt_value","pk"],"types":["","","","","",""],"values":[[0,"id","INTEGER",1,null,1],[1,"name","TEXT",0,null,0]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-func Test_LoadIntoMemory(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-
-	inmem, err := LoadIntoMemory(path, false)
-	if err != nil {
-		t.Fatalf("failed to create loaded in-memory database: %s", err.Error())
-	}
-
-	// Ensure it has been loaded correctly into the database
-	r, err = inmem.QueryStringStmt("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-func Test_DeserializeIntoMemory(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Transaction: true,
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(1, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(2, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(3, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(4, "fiona")`,
-			},
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert records: %s", err.Error())
-	}
-
-	// Get byte representation of database on disk which, according to SQLite docs
-	// is the same as a serialized version.
-	b, err := ioutil.ReadFile(path)
-	if err != nil {
-		t.Fatalf("failed to read database on disk: %s", err.Error())
-	}
-
-	newDB, err := DeserializeIntoMemory(b, false)
-	if err != nil {
-		t.Fatalf("failed to deserialize database: %s", err.Error())
-	}
-	defer newDB.Close()
-
-	ro, err := newDB.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"fiona"],[3,"fiona"],[4,"fiona"]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	// Write a lot of records to the new database, to ensure it's fully functional.
-	req = &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(name) VALUES("fiona")`,
-			},
-		},
-	}
-	for i := 0; i < 5000; i++ {
-		_, err = newDB.Execute(req, false)
-		if err != nil {
-			t.Fatalf("failed to insert records: %s", err.Error())
-		}
-	}
-	ro, err = newDB.QueryStringStmt(`SELECT COUNT(*) FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["COUNT(*)"],"types":[""],"values":[[5004]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_EmptyStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("")
-	if err != nil {
-		t.Fatalf("failed to execute empty statement: %s", err.Error())
-	}
-	_, err = db.ExecuteStringStmt(";")
-	if err != nil {
-		t.Fatalf("failed to execute empty statement with semicolon: %s", err.Error())
-	}
-}
-
-func Test_SimpleSingleStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("aoife")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"aoife"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.QueryStringStmt(`SELECT * FROM foo WHERE name="aoife"`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[2,"aoife"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.QueryStringStmt(`SELECT * FROM foo WHERE name="dana"`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.QueryStringStmt(`SELECT * FROM foo ORDER BY name`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[2,"aoife"],[1,"fiona"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.QueryStringStmt(`SELECT *,name FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name","name"],"types":["integer","text","text"],"values":[[1,"fiona","fiona"],[2,"aoife","aoife"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleSingleJSONStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (c0 VARCHAR(36), c1 JSON, c2 NCHAR, c3 NVARCHAR, c4 CLOB)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO foo(c0, c1, c2, c3, c4) VALUES("fiona", '{"mittens": "foobar"}', "bob", "dana", "declan")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["c0","c1","c2","c3","c4"],"types":["varchar(36)","json","nchar","nvarchar","clob"],"values":[["fiona","{\"mittens\": \"foobar\"}","bob","dana","declan"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-func Test_SimpleJoinStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE names (id INTEGER NOT NULL PRIMARY KEY, name TEXT, ssn TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO "names" VALUES(1,'bob','123-45-678')`,
-			},
-			{
-				Sql: `INSERT INTO "names" VALUES(2,'tom','111-22-333')`,
-			},
-			{
-				Sql: `INSERT INTO "names" VALUES(3,'matt','222-22-333')`,
-			},
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt("CREATE TABLE staff (id INTEGER NOT NULL PRIMARY KEY, employer TEXT, ssn TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO "staff" VALUES(1,'acme','222-22-333')`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt(`SELECT names.id,name,names.ssn,employer FROM names INNER JOIN staff ON staff.ssn = names.ssn`)
-	if err != nil {
-		t.Fatalf("failed to query table using JOIN: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name","ssn","employer"],"types":["integer","text","text","text"],"values":[[3,"matt","222-22-333","acme"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleSingleConcatStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt(`SELECT id || "_bar", name FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id || \"_bar\"","name"],"types":["","text"],"values":[["1_bar","fiona"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleMultiStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(name) VALUES("fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(name) VALUES("dana")`,
-			},
-		},
-	}
-	re, err := db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":1,"rows_affected":1},{"last_insert_id":2,"rows_affected":1}]`, asJSON(re); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	req = &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: `SELECT * FROM foo`,
-			},
-			{
-				Sql: `SELECT * FROM foo`,
-			},
-		},
-	}
-	ro, err := db.Query(req, false)
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"dana"]]},{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"dana"]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleSingleMultiLineStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: `
-CREATE TABLE foo (
-id INTEGER NOT NULL PRIMARY KEY,
-name TEXT
-)`,
-			},
-		},
-	}
-	_, err := db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req = &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(name) VALUES("fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(name) VALUES("dana")`,
-			},
-		},
-	}
-	re, err := db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":1,"rows_affected":1},{"last_insert_id":2,"rows_affected":1}]`, asJSON(re); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleFailingStatements_Execute(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	r, err := db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
-	if err != nil {
-		t.Fatalf("error executing insertion into non-existent table: %s", err.Error())
-	}
-	if exp, got := `[{"error":"no such table: foo"}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt(`CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)`)
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-	r, err = db.ExecuteStringStmt(`CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)`)
-	if err != nil {
-		t.Fatalf("failed to attempt creation of duplicate table: %s", err.Error())
-	}
-	if exp, got := `[{"error":"table foo already exists"}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt(`INSERT INTO foo(id, name) VALUES(11, "fiona")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":11,"rows_affected":1}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-	r, err = db.ExecuteStringStmt(`INSERT INTO foo(id, name) VALUES(11, "fiona")`)
-	if err != nil {
-		t.Fatalf("failed to attempt duplicate record insertion: %s", err.Error())
-	}
-	if exp, got := `[{"error":"UNIQUE constraint failed: foo.id"}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt(`utter nonsense`)
-	if err != nil {
-		if exp, got := `[{"error":"near \"utter\": syntax error"}]`, asJSON(r); exp != got {
-			t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-		}
-	}
-}
-
-func Test_SimpleFailingStatements_Query(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	ro, err := db.QueryStringStmt(`SELECT * FROM bar`)
-	if err != nil {
-		t.Fatalf("failed to attempt query of non-existent table: %s", err.Error())
-	}
-	if exp, got := `[{"error":"no such table: bar"}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	ro, err = db.QueryStringStmt(`SELECTxx * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to attempt nonsense query: %s", err.Error())
-	}
-	if exp, got := `[{"error":"near \"SELECTxx\": syntax error"}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-	r, err := db.QueryStringStmt(`utter nonsense`)
-	if err != nil {
-		if exp, got := `[{"error":"near \"utter\": syntax error"}]`, asJSON(r); exp != got {
-			t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-		}
-	}
-}
-
-func Test_SimplePragmaTableInfo(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	r, err := db.ExecuteStringStmt(`CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)`)
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	res, err := db.QueryStringStmt(`PRAGMA table_info("foo")`)
-	if err != nil {
-		t.Fatalf("failed to query a common table expression: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["cid","name","type","notnull","dflt_value","pk"],"types":["","","","","",""],"values":[[0,"id","INTEGER",1,null,1],[1,"name","TEXT",0,null,0]]}]`, asJSON(res); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_WriteOnQueryOnDiskDatabase(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	r, err := db.ExecuteStringStmt(`CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)`)
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt(`INSERT INTO foo(id, name) VALUES(1, "fiona")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":1,"rows_affected":1}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	ro, err := db.QueryStringStmt(`INSERT INTO foo(id, name) VALUES(2, "fiona")`)
-	if err != nil {
-		t.Fatalf("error attempting read-only write test: %s", err)
-	}
-	if exp, got := `[{"error":"attempt to change database via query operation"}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	ro, err = db.QueryStringStmt(`SELECT COUNT(*) FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["COUNT(*)"],"types":[""],"values":[[1]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_WriteOnQueryInMemDatabase(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	r, err := db.ExecuteStringStmt(`CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)`)
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt(`INSERT INTO foo(id, name) VALUES(1, "fiona")`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":1,"rows_affected":1}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	ro, err := db.QueryStringStmt(`INSERT INTO foo(id, name) VALUES(2, "fiona")`)
-	if err != nil {
-		t.Fatalf("error attempting read-only write test: %s", err)
-	}
-	if exp, got := `[{"error":"attempt to change database via query operation"}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-	ro, err = db.QueryStringStmt(`SELECT COUNT(*) FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["COUNT(*)"],"types":[""],"values":[[1]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_ConcurrentQueriesInMemory(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
 
 	r, err := db.ExecuteStringStmt(`CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)`)
 	if err != nil {
@@ -825,7 +245,7 @@ func Test_ConcurrentQueriesInMemory(t *testing.T) {
 			if err != nil {
 				t.Logf("failed to query table: %s", err.Error())
 			}
-			if exp, got := `[{"columns":["COUNT(*)"],"types":[""],"values":[[5000]]}]`, asJSON(ro); exp != got {
+			if exp, got := `[{"columns":["COUNT(*)"],"types":["integer"],"values":[[5000]]}]`, asJSON(ro); exp != got {
 				t.Logf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
 			}
 		}()
@@ -833,519 +253,8 @@ func Test_ConcurrentQueriesInMemory(t *testing.T) {
 	wg.Wait()
 }
 
-func Test_SimpleParameterizedStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: "INSERT INTO foo(name) VALUES(?)",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "fiona",
-						},
-					},
-				},
-			},
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	req.Statements[0].Parameters[0] = &command.Parameter{
-		Value: &command.Parameter_S{
-			S: "aoife",
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"aoife"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	req.Statements[0].Sql = "SELECT * FROM foo WHERE name=?"
-	req.Statements[0].Parameters[0] = &command.Parameter{
-		Value: &command.Parameter_S{
-			S: "aoife",
-		},
-	}
-	r, err = db.Query(req, false)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[2,"aoife"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	req.Statements[0].Parameters[0] = &command.Parameter{
-		Value: &command.Parameter_S{
-			S: "fiona",
-		},
-	}
-	r, err = db.Query(req, false)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	req = &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: "SELECT * FROM foo WHERE NAME=?",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "fiona",
-						},
-					},
-				},
-			},
-			{
-				Sql: "SELECT * FROM foo WHERE NAME=?",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "aoife",
-						},
-					},
-				},
-			},
-		},
-	}
-	r, err = db.Query(req, false)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]},{"columns":["id","name"],"types":["integer","text"],"values":[[2,"aoife"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleTwoParameterizedStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, first TEXT, last TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: "INSERT INTO foo(first, last) VALUES(?, ?)",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "bob",
-						},
-					},
-					{
-						Value: &command.Parameter_S{
-							S: "bobbers",
-						},
-					},
-				},
-			},
-		},
-	}
-
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","first","last"],"types":["integer","text","text"],"values":[[1,"bob","bobbers"]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleNilParameterizedStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, first TEXT, last TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: "INSERT INTO foo(first, last) VALUES(?, ?)",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "bob",
-						},
-					},
-					{
-						Value: nil,
-					},
-				},
-			},
-		},
-	}
-
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","first","last"],"types":["integer","text","text"],"values":[[1,"bob",null]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SimpleNamedParameterizedStatements(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, first TEXT, last TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	if _, err = db.ExecuteStringStmt(`INSERT INTO foo(first, last) VALUES("albert", "einstein")`); err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-	if _, err = db.ExecuteStringStmt(`INSERT INTO foo(first, last) VALUES("isaac", "newton")`); err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: "INSERT INTO foo(first, last) VALUES(:first, :last)",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "isaac",
-						},
-						Name: "first",
-					},
-					{
-						Value: &command.Parameter_S{
-							S: "asimov",
-						},
-						Name: "last",
-					},
-				},
-			},
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert parameterized statement: %s", err.Error())
-	}
-
-	rows, err := db.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","first","last"],"types":["integer","text","text"],"values":[[1,"albert","einstein"],[2,"isaac","newton"],[3,"isaac","asimov"]]}]`, asJSON(rows); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	req = &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: "SELECT * FROM foo WHERE first=:first AND last=:last",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "newton",
-						},
-						Name: "last",
-					},
-					{
-						Value: &command.Parameter_S{
-							S: "isaac",
-						},
-						Name: "first",
-					},
-				},
-			},
-		},
-	}
-	rn, err := db.Query(req, false)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","first","last"],"types":["integer","text","text"],"values":[[2,"isaac","newton"]]}]`, asJSON(rn); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	req = &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: "SELECT * FROM foo WHERE first=:first AND last=:last",
-				Parameters: []*command.Parameter{
-					{
-						Value: &command.Parameter_S{
-							S: "clarke",
-						},
-						Name: "last",
-					},
-					{
-						Value: &command.Parameter_S{
-							S: "isaac",
-						},
-						Name: "first",
-					},
-				},
-			},
-		},
-	}
-	rn, err = db.Query(req, false)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","first","last"],"types":["integer","text","text"]}]`, asJSON(rn); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_CommonTableExpressions(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE test(x foo)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO test VALUES(1)`)
-	if err != nil {
-		t.Fatalf("failed to insert record: %s", err.Error())
-	}
-
-	r, err := db.QueryStringStmt(`WITH bar AS (SELECT * FROM test) SELECT * FROM test WHERE x = 1`)
-	if err != nil {
-		t.Fatalf("failed to query a common table expression: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["x"],"types":["foo"],"values":[[1]]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-
-	r, err = db.QueryStringStmt(`WITH bar AS (SELECT * FROM test) SELECT * FROM test WHERE x = 2`)
-	if err != nil {
-		t.Fatalf("failed to query a common table expression: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["x"],"types":["foo"]}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_UniqueConstraints(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT, CONSTRAINT name_unique UNIQUE (name))")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	r, err := db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
-	if err != nil {
-		t.Fatalf("error executing insertion into table: %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":1,"rows_affected":1}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for INSERT\nexp: %s\ngot: %s", exp, got)
-	}
-
-	// UNIQUE constraint should fire.
-	r, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
-	if err != nil {
-		t.Fatalf("error executing insertion into table: %s", err.Error())
-	}
-	if exp, got := `[{"error":"UNIQUE constraint failed: foo.name"}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for INSERT\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-// Test_ConnectionIsolation test that ISOLATION behavior of on-disk databases doesn't
-// change unexpectedly.
-func Test_ConnectionIsolation(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	r, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt("BEGIN")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
-	if err != nil {
-		t.Fatalf("error executing insertion into table: %s", err.Error())
-	}
-
-	q, err := db.QueryStringStmt("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"]}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt("COMMIT")
-	if err != nil {
-		t.Fatalf("error executing insertion into table: %s", err.Error())
-	}
-
-	q, err = db.QueryStringStmt("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-// Test_ConnectionIsolationMemory test that ISOLATION behavior of in-memory databases doesn't
-// change unexpectedly.
-func Test_ConnectionIsolationMemory(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	r, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt("BEGIN")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-	if exp, got := `[{}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-
-	r, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
-	if err != nil {
-		t.Fatalf("error executing insertion into table: %s", err.Error())
-	}
-
-	_, err = db.QueryStringStmt("SELECT * FROM foo")
-	if err == nil {
-		t.Fatalf("SELECT should have received error")
-	}
-	if exp, got := "database is locked", err.Error(); exp != got {
-		t.Fatalf("error incorrect, exp %s, got %s", exp, got)
-	}
-}
-
-func Test_Size(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	if _, err := db.Size(); err != nil {
-		t.Fatalf("failed to read database size: %s", err)
-	}
-}
-
-func Test_DBFileSize(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	if _, err := db.FileSize(); err != nil {
-		t.Fatalf("failed to read database file size: %s", err)
-	}
-}
-
-func Test_PartialFail(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(1, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(2, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(1, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(4, "fiona")`,
-			},
-		},
-	}
-	r, err := db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert records: %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":1,"rows_affected":1},{"last_insert_id":2,"rows_affected":1},{"error":"UNIQUE constraint failed: foo.id"},{"last_insert_id":4,"rows_affected":1}]`, asJSON(r); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-	ro, err := db.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"fiona"],[4,"fiona"]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
 func Test_SimpleTransaction(t *testing.T) {
-	db, path := mustCreateDatabase()
+	db, path := mustCreateOnDiskDatabaseWAL()
 	defer db.Close()
 	defer os.Remove(path)
 
@@ -1388,7 +297,7 @@ func Test_SimpleTransaction(t *testing.T) {
 }
 
 func Test_PartialFailTransaction(t *testing.T) {
-	db, path := mustCreateDatabase()
+	db, path := mustCreateOnDiskDatabaseWAL()
 	defer db.Close()
 	defer os.Remove(path)
 
@@ -1430,326 +339,644 @@ func Test_PartialFailTransaction(t *testing.T) {
 	}
 }
 
-func Test_Backup(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
+func Test_IsValidSQLiteOnDisk(t *testing.T) {
+	path := mustTempFile()
 	defer os.Remove(path)
 
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
+	dsn := fmt.Sprintf("file:%s", path)
+	db, err := sql.Open("sqlite3", dsn)
+	if err != nil {
+		t.Fatalf("failed to create SQLite database: %s", err.Error())
+	}
+	_, err = db.Exec("CREATE TABLE foo (name TEXT)")
 	if err != nil {
 		t.Fatalf("failed to create table: %s", err.Error())
 	}
-
-	req := &command.Request{
-		Transaction: true,
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(1, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(2, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(3, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(4, "fiona")`,
-			},
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert records: %s", err.Error())
+	if err := db.Close(); err != nil {
+		t.Fatalf("failed to close database: %s", err.Error())
 	}
 
-	dstDB := mustTempFile()
-	defer os.Remove(dstDB)
-
-	err = db.Backup(dstDB)
-	if err != nil {
-		t.Fatalf("failed to backup database: %s", err.Error())
+	if !IsValidSQLiteFile(path) {
+		t.Fatalf("good SQLite file marked as invalid")
 	}
 
-	newDB, err := Open(dstDB, false)
+	data, err := os.ReadFile(path)
 	if err != nil {
-		t.Fatalf("failed to open backup database: %s", err.Error())
+		t.Fatalf("failed to read SQLite file: %s", err.Error())
 	}
-	defer newDB.Close()
-	defer os.Remove(dstDB)
-	ro, err := newDB.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"fiona"],[3,"fiona"],[4,"fiona"]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
+	if !IsValidSQLiteData(data) {
+		t.Fatalf("good SQLite data marked as invalid")
 	}
 }
 
-func Test_Copy(t *testing.T) {
-	srcDB, path := mustCreateDatabase()
-	defer srcDB.Close()
+func Test_CheckIntegrityOnDisk(t *testing.T) {
+	path := mustTempFile()
 	defer os.Remove(path)
 
-	_, err := srcDB.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
+	dsn := fmt.Sprintf("file:%s", path)
+	db, err := sql.Open("sqlite3", dsn)
+	if err != nil {
+		t.Fatalf("failed to create SQLite database: %s", err.Error())
+	}
+	_, err = db.Exec("CREATE TABLE foo (name TEXT)")
 	if err != nil {
 		t.Fatalf("failed to create table: %s", err.Error())
 	}
-
-	req := &command.Request{
-		Transaction: true,
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(1, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(2, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(3, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(4, "fiona")`,
-			},
-		},
-	}
-	_, err = srcDB.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert records: %s", err.Error())
+	if err := db.Close(); err != nil {
+		t.Fatalf("failed to close database: %s", err.Error())
 	}
 
-	dstFile := mustTempFile()
-	defer os.Remove(dstFile)
-	dstDB, err := Open(dstFile, false)
-	if err != nil {
-		t.Fatalf("failed to open destination database: %s", err)
-	}
-	defer dstDB.Close()
-
-	err = srcDB.Copy(dstDB)
-	if err != nil {
-		t.Fatalf("failed to copy database: %s", err.Error())
-	}
-
-	ro, err := dstDB.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"fiona"],[3,"fiona"],[4,"fiona"]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SerializeOnDisk(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Transaction: true,
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(1, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(2, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(3, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(4, "fiona")`,
-			},
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert records: %s", err.Error())
-	}
-
-	dstDB, err := ioutil.TempFile("", "rqlite-bak-")
-	if err != nil {
-		t.Fatalf("failed to create temp file: %s", err.Error())
-	}
-	dstDB.Close()
-	defer os.Remove(dstDB.Name())
-
-	// Get the bytes, and write to a temp file.
-	b, err := db.Serialize()
-	if err != nil {
-		t.Fatalf("failed to serialize database: %s", err.Error())
-	}
-	err = ioutil.WriteFile(dstDB.Name(), b, 0644)
-	if err != nil {
-		t.Fatalf("failed to write serialized database to file: %s", err.Error())
-	}
-
-	newDB, err := Open(dstDB.Name(), false)
-	if err != nil {
-		t.Fatalf("failed to open on-disk serialized database: %s", err.Error())
-	}
-	defer newDB.Close()
-	defer os.Remove(dstDB.Name())
-	ro, err := newDB.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"fiona"],[3,"fiona"],[4,"fiona"]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_SerializeInMemory(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	req := &command.Request{
-		Transaction: true,
-		Statements: []*command.Statement{
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(1, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(2, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(3, "fiona")`,
-			},
-			{
-				Sql: `INSERT INTO foo(id, name) VALUES(4, "fiona")`,
-			},
-		},
-	}
-	_, err = db.Execute(req, false)
-	if err != nil {
-		t.Fatalf("failed to insert records: %s", err.Error())
-	}
-
-	dstDB, err := ioutil.TempFile("", "rqlite-bak-")
-	if err != nil {
-		t.Fatalf("failed to create temp file: %s", err.Error())
-	}
-	dstDB.Close()
-	defer os.Remove(dstDB.Name())
-
-	// Get the bytes, and write to a temp file.
-	b, err := db.Serialize()
-	if err != nil {
-		t.Fatalf("failed to serialize database: %s", err.Error())
-	}
-	err = ioutil.WriteFile(dstDB.Name(), b, 0644)
-	if err != nil {
-		t.Fatalf("failed to write serialized database to file: %s", err.Error())
-	}
-
-	newDB, err := Open(dstDB.Name(), false)
-	if err != nil {
-		t.Fatalf("failed to open on-disk serialized database: %s", err.Error())
-	}
-	defer newDB.Close()
-	defer os.Remove(dstDB.Name())
-	ro, err := newDB.QueryStringStmt(`SELECT * FROM foo`)
-	if err != nil {
-		t.Fatalf("failed to query table: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"fiona"],[3,"fiona"],[4,"fiona"]]}]`, asJSON(ro); exp != got {
-		t.Fatalf("unexpected results for query\nexp: %s\ngot: %s", exp, got)
-	}
-}
-
-func Test_Dump(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	_, err := db.ExecuteStringStmt(chinook.DB)
-	if err != nil {
-		t.Fatalf("failed to load chinook dump: %s", err.Error())
-	}
-
-	var b strings.Builder
-	if err := db.Dump(&b); err != nil {
-		t.Fatalf("failed to dump database: %s", err.Error())
-	}
-
-	if b.String() != chinook.DB {
-		t.Fatal("dumped database does not equal entered database")
-	}
-}
-
-func Test_DumpMemory(t *testing.T) {
-	db, path := mustCreateDatabase()
-	defer db.Close()
-	defer os.Remove(path)
-
-	inmem, err := LoadIntoMemory(path, false)
-	if err != nil {
-		t.Fatalf("failed to create loaded in-memory database: %s", err.Error())
-	}
-
-	_, err = inmem.ExecuteStringStmt(chinook.DB)
-	if err != nil {
-		t.Fatalf("failed to load chinook dump: %s", err.Error())
-	}
-
-	var b strings.Builder
-	if err := inmem.Dump(&b); err != nil {
-		t.Fatalf("failed to dump database: %s", err.Error())
-	}
-
-	if b.String() != chinook.DB {
-		t.Fatal("dumped database does not equal entered database")
-	}
-}
-
-// Test_1GiBInMemory tests that in-memory databases larger than 1GiB,
-// but smaller than 2GiB, can be created without issue.
-func Test_1GiBInMemory(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, txt TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	stmt := fmt.Sprintf(`INSERT INTO foo(txt) VALUES("%s")`, text.Lorum)
-	for i := 0; i < 1715017; i++ {
-		r, err := db.ExecuteStringStmt(stmt)
+	for _, check := range []bool{true, false} {
+		ok, err := CheckIntegrity(path, check)
 		if err != nil {
-			t.Fatalf("failed to Execute statement %s", err.Error())
+			t.Fatalf("failed to check integrity of database (full=%t): %s", check, err.Error())
 		}
-		if len(r) != 1 {
-			t.Fatalf("unexpected length for Execute results: %d", len(r))
-		}
-		if r[0].GetError() != "" {
-			t.Fatalf("failed to insert record: %s", r[0].GetError())
+		if !ok {
+			t.Fatalf("database failed integrity check (full=%t)", check)
 		}
 	}
 
-	r, err := db.ExecuteStringStmt(stmt)
+	sz := int(mustFileSize(path))
+	b, err := os.ReadFile(path)
 	if err != nil {
-		t.Fatalf("failed to insert record %s", err.Error())
-	}
-	if exp, got := `[{"last_insert_id":1715018,"rows_affected":1}]`, asJSON(r); exp != got {
-		t.Fatalf("got incorrect response, exp: %s, got: %s", exp, got)
+		t.Fatalf("failed to read SQLite file: %s", err.Error())
 	}
 
-	sz, err := db.Size()
-	if err != nil {
-		t.Fatalf("failed to get size: %s", err.Error())
+	// Chop the file in half, integrity check shouldn't even be error-free
+	if err := os.WriteFile(path, b[:len(b)-sz/2], 0644); err != nil {
+		t.Fatalf("failed to write SQLite file: %s", err.Error())
 	}
-	if sz <= 1024*1024*1024 {
-		t.Fatalf("failed to create a database greater than 1 GiB in size: %d", sz)
+	for _, check := range []bool{true, false} {
+		_, err := CheckIntegrity(path, check)
+		if err == nil {
+			t.Fatalf("succeeded checking integrity of database (full=%t): %s", check, err.Error())
+		}
+	}
+	// Unable to create a data set that actually fails integrity check.
+}
+
+func Test_IsWALModeEnabledOnDiskDELETE(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+
+	dsn := fmt.Sprintf("file:%s", path)
+	db, err := sql.Open("sqlite3", dsn)
+	if err != nil {
+		t.Fatalf("failed to create SQLite database: %s", err.Error())
+	}
+	_, err = db.Exec("CREATE TABLE foo (name TEXT)")
+	if err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	if err := db.Close(); err != nil {
+		t.Fatalf("failed to close database: %s", err.Error())
+	}
+
+	if !IsDELETEModeEnabledSQLiteFile(path) {
+		t.Fatalf("DELETE file marked as non-DELETE")
+	}
+	if IsWALModeEnabledSQLiteFile(path) {
+		t.Fatalf("non WAL file marked as WAL")
+	}
+
+	data, err := os.ReadFile(path)
+	if err != nil {
+		t.Fatalf("failed to read SQLite file: %s", err.Error())
+	}
+	if IsWALModeEnabled(data) {
+		t.Fatalf("non WAL data marked as WAL")
+	}
+	if !IsDELETEModeEnabled(data) {
+		t.Fatalf("data marked as non-DELETE")
+	}
+}
+
+func Test_IsWALModeEnabledOnDiskWAL(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+
+	dsn := fmt.Sprintf("file:%s", path)
+	db, err := sql.Open("sqlite3", dsn)
+	if err != nil {
+		t.Fatalf("failed to create SQLite database: %s", err.Error())
+	}
+	_, err = db.Exec("CREATE TABLE foo (name TEXT)")
+	if err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	_, err = db.Exec("PRAGMA journal_mode=WAL")
+	if err != nil {
+		t.Fatalf("failed to enable WAL mode: %s", err.Error())
+	}
+
+	_, err = db.Exec(`INSERT INTO foo(name) VALUES("fiona")`)
+	if err != nil {
+		t.Fatalf("error inserting record into table: %s", err.Error())
+	}
+
+	if !IsValidSQLiteWALFile(path + "-wal") {
+		t.Fatalf("WAL file marked not marked as valid")
+	}
+
+	if err := db.Close(); err != nil {
+		t.Fatalf("failed to close database: %s", err.Error())
+	}
+
+	if !IsWALModeEnabledSQLiteFile(path) {
+		t.Fatalf("WAL file marked as non-WAL")
+	}
+	if IsDELETEModeEnabledSQLiteFile(path) {
+		t.Fatalf("WAL file marked as DELETE")
+	}
+
+	data, err := os.ReadFile(path)
+	if err != nil {
+		t.Fatalf("failed to read SQLite file: %s", err.Error())
+	}
+	if !IsWALModeEnabled(data) {
+		t.Fatalf("WAL data marked as non-WAL")
+	}
+	if IsDELETEModeEnabled(data) {
+		t.Fatalf("WAL data marked as DELETE")
+	}
+}
+func Test_EnsureDelete(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+
+	db, err := Open(path, false, true)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	defer db.Close()
+	if !IsWALModeEnabledSQLiteFile(path) {
+		t.Fatalf("WAL file marked as non-WAL")
+	}
+
+	if err := EnsureDeleteMode(path); err != nil {
+		t.Fatalf("failed to ensure DELETE mode: %s", err.Error())
+	}
+	if !IsDELETEModeEnabledSQLiteFile(path) {
+		t.Fatalf("database not marked as DELETE mode")
+	}
+}
+
+func Test_DBLastModified(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+	db, err := Open(path, false, true)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	defer db.Close()
+
+	lm, err := db.LastModified()
+	if err != nil {
+		t.Fatalf("failed to get last modified time: %s", err.Error())
+	}
+	if lm.IsZero() {
+		t.Fatalf("last modified time is zero")
+	}
+
+	// Write some data, check time is later.
+	_, err = db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
+	if err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+
+	lm2, err := db.LastModified()
+	if err != nil {
+		t.Fatalf("failed to get last modified time: %s", err.Error())
+	}
+	if lm2.Before(lm) {
+		t.Fatalf("last modified time not updated")
+	}
+
+	// Checkpoint, check time is later.
+	if err := db.Checkpoint(); err != nil {
+		t.Fatalf("failed to checkpoint database: %s", err.Error())
+	}
+	lm3, err := db.LastModified()
+	if err != nil {
+		t.Fatalf("failed to get last modified time: %s", err.Error())
+	}
+	if lm3.Before(lm2) {
+		t.Fatalf("last modified time not updated after checkpoint")
+	}
+
+	// Call again, without changes, check time is same.
+	lm4, err := db.LastModified()
+	if err != nil {
+		t.Fatalf("failed to get last modified time: %s", err.Error())
+	}
+	if !lm4.Equal(lm3) {
+		t.Fatalf("last modified time updated without changes")
+	}
+}
+
+// Test_WALDatabaseCreatedOK tests that a WAL file is created, and that
+// a checkpoint succeeds
+func Test_WALDatabaseCreatedOK(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+
+	db, err := Open(path, false, true)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	defer db.Close()
+
+	if !db.WALEnabled() {
+		t.Fatalf("WAL mode not enabled")
+	}
+
+	if _, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)"); err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+
+	if !IsWALModeEnabledSQLiteFile(path) {
+		t.Fatalf("SQLite file not marked as WAL")
+	}
+
+	walPath := path + "-wal"
+	if _, err := os.Stat(walPath); os.IsNotExist(err) {
+		t.Fatalf("WAL file does not exist")
+	}
+
+	if err := db.Checkpoint(); err != nil {
+		t.Fatalf("failed to checkpoint database in WAL mode: %s", err.Error())
+	}
+	// Checkpoint a second time, to ensure it's idempotent.
+	if err := db.Checkpoint(); err != nil {
+		t.Fatalf("failed to checkpoint database in WAL mode: %s", err.Error())
+	}
+}
+
+// Test_WALDatabaseCheckpointOKNoWAL tests that a checkpoint succeeds
+// even when no WAL file exists.
+func Test_WALDatabaseCheckpointOKNoWAL(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+
+	db, err := Open(path, false, true)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	if !db.WALEnabled() {
+		t.Fatalf("WAL mode not enabled")
+	}
+	if fileExists(db.WALPath()) {
+		t.Fatalf("WAL file exists when no writes have happened")
+	}
+	defer db.Close()
+	if err := db.Checkpoint(); err != nil {
+		t.Fatalf("failed to checkpoint database in WAL mode with non-existent WAL: %s", err.Error())
+	}
+}
+
+// Test_WALDatabaseCheckpointOKDelete tests that a checkpoint returns no error
+// even when the database is opened in DELETE mode.
+func Test_WALDatabaseCheckpointOKDelete(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+
+	db, err := Open(path, false, false)
+	if err != nil {
+		t.Fatalf("failed to open database in DELETE mode: %s", err.Error())
+	}
+	if db.WALEnabled() {
+		t.Fatalf("WAL mode enabled")
+	}
+	defer db.Close()
+	if err := db.Checkpoint(); err != nil {
+		t.Fatalf("failed to checkpoint database in DELETE mode: %s", err.Error())
+	}
+}
+
+// Test_WALDatabaseCreatedOKFromDELETE tests that a WAL database is created properly,
+// even when supplied with a DELETE-mode database.
+func Test_WALDatabaseCreatedOKFromDELETE(t *testing.T) {
+	deletePath := mustTempFile()
+	defer os.Remove(deletePath)
+	deleteDB, err := Open(deletePath, false, false)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	defer deleteDB.Close()
+	if _, err := deleteDB.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)"); err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	_, err = deleteDB.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
+	if err != nil {
+		t.Fatalf("error executing insertion into table: %s", err.Error())
+	}
+
+	walDB, err := Open(deletePath, false, true)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	defer walDB.Close()
+	if !IsWALModeEnabledSQLiteFile(deletePath) {
+		t.Fatalf("SQLite file not marked as WAL")
+	}
+	rows, err := walDB.QueryStringStmt("SELECT * FROM foo")
+	if err != nil {
+		t.Fatalf("failed to query WAL table: %s", err.Error())
+	}
+	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]`, asJSON(rows); exp != got {
+		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+	}
+}
+
+// Test_DELETEDatabaseCreatedOKFromWAL tests that a DELETE database is created properly,
+// even when supplied with a WAL-mode database.
+func Test_DELETEDatabaseCreatedOKFromWAL(t *testing.T) {
+	walPath := mustTempFile()
+	defer os.Remove(walPath)
+	walDB, err := Open(walPath, false, true)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	defer walDB.Close()
+	if _, err := walDB.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)"); err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	_, err = walDB.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
+	if err != nil {
+		t.Fatalf("error executing insertion into table: %s", err.Error())
+	}
+	if err := walDB.Close(); err != nil {
+		// Closing the WAL database is required if it's to be opened in DELETE mode.
+		t.Fatalf("failed to close database: %s", err.Error())
+	}
+
+	deleteDB, err2 := Open(walPath, false, false)
+	if err2 != nil {
+		t.Fatalf("failed to open database in DELETE mode: %s", err2.Error())
+	}
+	defer deleteDB.Close()
+	if !IsDELETEModeEnabledSQLiteFile(walPath) {
+		t.Fatalf("SQLite file not marked as WAL")
+	}
+	rows, err := deleteDB.QueryStringStmt("SELECT * FROM foo")
+	if err != nil {
+		t.Fatalf("failed to query WAL table: %s", err.Error())
+	}
+	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]`, asJSON(rows); exp != got {
+		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+	}
+}
+
+func Test_WALDisableCheckpointing(t *testing.T) {
+	path := mustTempFile()
+	defer os.Remove(path)
+
+	db, err := Open(path, false, true)
+	if err != nil {
+		t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+	}
+	defer db.Close()
+	if !db.WALEnabled() {
+		t.Fatalf("WAL mode not enabled")
+	}
+
+	// Test that databases open with checkpoint disabled by default.
+	// This is critical.
+	n, err := db.GetCheckpointing()
+	if err != nil {
+		t.Fatalf("failed to get checkpoint value: %s", err.Error())
+	}
+	if exp, got := 0, n; exp != got {
+		t.Fatalf("unexpected checkpoint value, expected %d, got %d", exp, got)
+	}
+
+	if err := db.EnableCheckpointing(); err != nil {
+		t.Fatalf("failed to disable checkpointing: %s", err.Error())
+	}
+	n, err = db.GetCheckpointing()
+	if err != nil {
+		t.Fatalf("failed to get checkpoint value: %s", err.Error())
+	}
+	if exp, got := 1000, n; exp != got {
+		t.Fatalf("unexpected checkpoint value, expected %d, got %d", exp, got)
+	}
+}
+
+// Test_WALReplayOK tests that WAL files are replayed as expected.
+func Test_WALReplayOK(t *testing.T) {
+	testFunc := func(t *testing.T, replayIntoDelete bool) {
+		dbPath := mustTempFile()
+		defer os.Remove(dbPath)
+		db, err := Open(dbPath, false, true)
+		if err != nil {
+			t.Fatalf("failed to open database in WAL mode: %s", err.Error())
+		}
+		defer db.Close()
+
+		n, err := db.GetCheckpointing()
+		if err != nil {
+			t.Fatalf("failed to get checkpoint value: %s", err.Error())
+		}
+		if exp, got := 0, n; exp != got {
+			t.Fatalf("unexpected checkpoint value, expected %d, got %d", exp, got)
+		}
+
+		dbFile := filepath.Base(dbPath)
+		walPath := dbPath + "-wal"
+		walFile := filepath.Base(walPath)
+
+		replayDir := mustTempDir()
+		defer os.RemoveAll(replayDir)
+		replayDBPath := filepath.Join(replayDir, dbFile)
+
+		// Create and copy the SQLite file and WAL #1
+		if _, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)"); err != nil {
+			t.Fatalf("failed to create table: %s", err.Error())
+		}
+		if !fileExists(walPath) {
+			t.Fatalf("WAL file at %s does not exist", walPath)
+		}
+		mustCopyFile(replayDBPath, dbPath)
+		mustCopyFile(filepath.Join(replayDir, walFile+"_001"), walPath)
+		if err := db.Checkpoint(); err != nil {
+			t.Fatalf("failed to checkpoint database in WAL mode: %s", err.Error())
+		}
+
+		// Create and copy WAL #2
+		_, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
+		if err != nil {
+			t.Fatalf("error executing insertion into table: %s", err.Error())
+		}
+		if !fileExists(walPath) {
+			t.Fatalf("WAL file at %s does not exist", walPath)
+		}
+		mustCopyFile(filepath.Join(replayDir, walFile+"_002"), walPath)
+		if err := db.Checkpoint(); err != nil {
+			t.Fatalf("failed to checkpoint database in WAL mode: %s", err.Error())
+		}
+
+		// Create and copy WAL #3
+		_, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("declan")`)
+		if err != nil {
+			t.Fatalf("error executing insertion into table: %s", err.Error())
+		}
+		if !fileExists(walPath) {
+			t.Fatalf("WAL file at %s does not exist", walPath)
+		}
+		mustCopyFile(filepath.Join(replayDir, walFile+"_003"), walPath)
+
+		if err := db.Close(); err != nil {
+			t.Fatalf("failed to close database: %s", err.Error())
+		}
+
+		wals := []string{
+			filepath.Join(replayDir, walFile+"_001"),
+			filepath.Join(replayDir, walFile+"_002"),
+			filepath.Join(replayDir, walFile+"_003"),
+		}
+		if err := ReplayWAL(replayDBPath, wals, replayIntoDelete); err != nil {
+			t.Fatalf("failed to replay WAL files: %s", err.Error())
+		}
+
+		if replayIntoDelete {
+			if !IsDELETEModeEnabledSQLiteFile(replayDBPath) {
+				t.Fatal("replayed database not marked as DELETE mode")
+			}
+		} else {
+			if !IsWALModeEnabledSQLiteFile(replayDBPath) {
+				t.Fatal("replayed database not marked as WAL mode")
+			}
+		}
+
+		// Check that there are no files related to WALs in the replay directory
+		// Both the copied WAL files should be gone, and there should be no
+		// "real" WAL file either.
+		walFiles, err := filepath.Glob(filepath.Join(replayDir, "*-wal*"))
+		if err != nil {
+			t.Fatalf("failed to glob replay directory: %s", err.Error())
+		}
+		if len(walFiles) != 0 {
+			t.Fatalf("replay directory contains WAL files: %s", walFiles)
+		}
+
+		replayedDB, err := Open(replayDBPath, false, true)
+		if err != nil {
+			t.Fatalf("failed to open replayed database: %s", err.Error())
+		}
+		rows, err := replayedDB.QueryStringStmt("SELECT * FROM foo")
+		if err != nil {
+			t.Fatalf("failed to query WAL table: %s", err.Error())
+		}
+		if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"],[2,"declan"]]}]`, asJSON(rows); exp != got {
+			t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+		}
+	}
+
+	t.Run("replayIntoWAL", func(t *testing.T) {
+		testFunc(t, false)
+	})
+	t.Run("replayIntoDELETE", func(t *testing.T) {
+		testFunc(t, true)
+	})
+}
+
+func Test_WALReplayFailures(t *testing.T) {
+	dbDir := mustTempDir()
+	defer os.RemoveAll(dbDir)
+	walDir := mustTempDir()
+	defer os.RemoveAll(walDir)
+
+	err := ReplayWAL(filepath.Join(dbDir, "foo.db"), []string{filepath.Join(walDir, "foo.db-wal")}, false)
+	if err != ErrWALReplayDirectoryMismatch {
+		t.Fatalf("expected %s, got %s", ErrWALReplayDirectoryMismatch, err.Error())
+	}
+}
+
+func test_FileCreationOnDisk(t *testing.T, db *DB) {
+	defer db.Close()
+	if db.FKEnabled() {
+		t.Fatal("FK constraints marked as enabled")
+	}
+
+	// Confirm checkpoint works on all types of on-disk databases. Worst case, this
+	// should be ignored.
+	if err := db.Checkpoint(); err != nil {
+		t.Fatalf("failed to checkpoint database in DELETE mode: %s", err.Error())
+	}
+}
+
+// test_ConnectionIsolationOnDisk test that ISOLATION behavior of on-disk databases doesn't
+// change unexpectedly.
+func test_ConnectionIsolationOnDisk(t *testing.T, db *DB) {
+	r, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
+	if err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	if exp, got := `[{}]`, asJSON(r); exp != got {
+		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+	}
+
+	r, err = db.ExecuteStringStmt("BEGIN")
+	if err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	if exp, got := `[{}]`, asJSON(r); exp != got {
+		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+	}
+
+	r, err = db.ExecuteStringStmt(`INSERT INTO foo(name) VALUES("fiona")`)
+	if err != nil {
+		t.Fatalf("error executing insertion into table: %s", err.Error())
+	}
+	if exp, got := `[{"last_insert_id":1,"rows_affected":1}]`, asJSON(r); exp != got {
+		t.Fatalf("unexpected results for execute, expected %s, got %s", exp, got)
+	}
+
+	q, err := db.QueryStringStmt("SELECT * FROM foo")
+	if err != nil {
+		t.Fatalf("failed to query empty table: %s", err.Error())
+	}
+	if exp, got := `[{"columns":["id","name"],"types":["integer","text"]}]`, asJSON(q); exp != got {
+		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+	}
+
+	_, err = db.ExecuteStringStmt("COMMIT")
+	if err != nil {
+		t.Fatalf("error executing insertion into table: %s", err.Error())
+	}
+
+	q, err = db.QueryStringStmt("SELECT * FROM foo")
+	if err != nil {
+		t.Fatalf("failed to query empty table: %s", err.Error())
+	}
+	if exp, got := `[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]`, asJSON(q); exp != got {
+		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
+	}
+}
+
+func Test_DatabaseCommonOnDiskOperations(t *testing.T) {
+	testCases := []struct {
+		name     string
+		testFunc func(*testing.T, *DB)
+	}{
+		{"FileCreationOnDisk", test_FileCreationOnDisk},
+		{"ConnectionIsolationOnDisk", test_ConnectionIsolationOnDisk},
+	}
+
+	for _, tc := range testCases {
+		db, path := mustCreateOnDiskDatabase()
+		defer db.Close()
+		defer os.Remove(path)
+		t.Run(tc.name+":disk", func(t *testing.T) {
+			tc.testFunc(t, db)
+		})
+
+		db, path = mustCreateOnDiskDatabaseWAL()
+		defer db.Close()
+		defer os.Remove(path)
+		t.Run(tc.name+":wal", func(t *testing.T) {
+			tc.testFunc(t, db)
+		})
 	}
 }
 
@@ -1759,8 +986,9 @@ func Test_1GiBInMemory(t *testing.T) {
 //
 // See https://github.com/mattn/go-sqlite3/issues/959#issuecomment-890283264
 func Test_ParallelOperationsInMemory(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
+	db, path := mustCreateOnDiskDatabaseWAL()
 	defer db.Close()
+	defer os.Remove(path)
 
 	if _, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)"); err != nil {
 		t.Fatalf("failed to create table: %s", err.Error())
@@ -1860,140 +1088,32 @@ func Test_ParallelOperationsInMemory(t *testing.T) {
 	exWg.Wait()
 }
 
-func Test_DBSTAT_table(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE foo (id INTEGER NOT NULL PRIMARY KEY, name TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	q, err := db.QueryStringStmt("SELECT * FROM dbstat")
-	if err != nil {
-		t.Fatalf("failed to query empty table: %s", err.Error())
-	}
-	if exp, got := `["name","path","pageno","pagetype","ncell","payload","unused","mx_payload","pgoffset","pgsize"]`, asJSON(q[0].Columns); exp != got {
-		t.Fatalf("unexpected results for query, expected %s, got %s", exp, got)
-	}
-}
-
-func Test_JSON1(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE customer(name,phone)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	_, err = db.ExecuteStringStmt(`INSERT INTO customer (name, phone) VALUES("fiona", json('{"mobile":"789111", "home":"123456"}'))`)
-	if err != nil {
-		t.Fatalf("failed to insert JSON record: %s", err.Error())
-	}
-
-	q, err := db.QueryStringStmt("SELECT customer.phone FROM customer WHERE customer.name=='fiona'")
-	if err != nil {
-		t.Fatalf("failed to perform simple SELECT: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["phone"],"types":[""],"values":[["{\"mobile\":\"789111\",\"home\":\"123456\"}"]]}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for simple query, expected %s, got %s", exp, got)
-	}
-	q, err = db.QueryStringStmt("SELECT json_extract(customer.phone, '$.mobile') FROM customer")
-	if err != nil {
-		t.Fatalf("failed to perform simple SELECT: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["json_extract(customer.phone, '$.mobile')"],"types":[""],"values":[["789111"]]}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for JSON query, expected %s, got %s", exp, got)
-	}
-	q, err = db.QueryStringStmt("SELECT customer.phone ->> '$.mobile' FROM customer")
-	if err != nil {
-		t.Fatalf("failed to perform simple SELECT: %s", err.Error())
-	}
-	if exp, got := `[{"columns":["customer.phone ->> '$.mobile'"],"types":[""],"values":[["789111"]]}]`, asJSON(q); exp != got {
-		t.Fatalf("unexpected results for JSON query, expected %s, got %s", exp, got)
-	}
-}
-
-// Test_TableCreationInMemoryLoadRaw tests for https://sqlite.org/forum/forumpost/d443fb0730
-func Test_TableCreationInMemoryLoadRaw(t *testing.T) {
-	db := mustCreateInMemoryDatabase()
-	defer db.Close()
-
-	_, err := db.ExecuteStringStmt("CREATE TABLE logs (entry TEXT)")
-	if err != nil {
-		t.Fatalf("failed to create table: %s", err.Error())
-	}
-
-	done := make(chan struct{})
-	defer close(done)
-
-	// Insert some records continually, as fast as possible. Do it from a goroutine.
-	go func() {
-		for {
-			select {
-			case <-done:
-				return
-			default:
-				_, err := db.ExecuteStringStmt(`INSERT INTO logs(entry) VALUES("hello")`)
-				if err != nil {
-					return
-				}
-			}
-		}
-	}()
-
-	// Get the count over and over again.
-	for i := 0; i < 5000; i++ {
-		rows, err := db.QueryStringStmt(`SELECT COUNT(*) FROM logs`)
-		if err != nil {
-			t.Fatalf("failed to query for count: %s", err)
-		}
-
-		if rows[0].Error != "" {
-			t.Fatalf("rows had error after %d queries: %s", i, rows[0].Error)
-		}
-	}
-}
-
-func mustCreateDatabase() (*DB, string) {
+func mustCreateOnDiskDatabase() (*DB, string) {
 	var err error
 	f := mustTempFile()
-	db, err := Open(f, false)
+	db, err := Open(f, false, false)
 	if err != nil {
-		panic("failed to open database")
+		panic("failed to open database in DELETE mode")
 	}
-
 	return db, f
 }
 
-func mustCreateInMemoryDatabase() *DB {
-	db, err := OpenInMemory(false)
-	if err != nil {
-		panic("failed to open in-memory database")
-	}
-	return db
-}
-
-func mustCreateInMemoryDatabaseFK() *DB {
-	db, err := OpenInMemory(true)
-	if err != nil {
-		panic("failed to open in-memory database with foreign key constraints")
-	}
-	return db
-}
-
-func mustWriteAndOpenDatabase(b []byte) (*DB, string) {
+func mustCreateOnDiskDatabaseWAL() (*DB, string) {
 	var err error
 	f := mustTempFile()
-	err = ioutil.WriteFile(f, b, 0660)
+	db, err := Open(f, false, true)
 	if err != nil {
-		panic("failed to write file")
+		panic("failed to open database in WAL mode")
 	}
+	return db, f
+}
 
-	db, err := Open(f, false)
+func mustCreateOnDiskDatabaseWALFK() (*DB, string) {
+	var err error
+	f := mustTempFile()
+	db, err := Open(f, true, true)
 	if err != nil {
-		panic("failed to open database")
+		panic("failed to open database in WAL mode")
 	}
 	return db, f
 }
@@ -2001,18 +1121,12 @@ func mustWriteAndOpenDatabase(b []byte) (*DB, string) {
 // mustExecute executes a statement, and panics on failure. Used for statements
 // that should never fail, even taking into account test setup.
 func mustExecute(db *DB, stmt string) {
-	_, err := db.ExecuteStringStmt(stmt)
+	r, err := db.ExecuteStringStmt(stmt)
 	if err != nil {
 		panic(fmt.Sprintf("failed to execute statement: %s", err.Error()))
 	}
-}
-
-// mustQuery executes a statement, and panics on failure. Used for statements
-// that should never fail, even taking into account test setup.
-func mustQuery(db *DB, stmt string) {
-	_, err := db.QueryStringStmt(stmt)
-	if err != nil {
-		panic(fmt.Sprintf("failed to query: %s", err.Error()))
+	if r[0].Error != "" {
+		panic(fmt.Sprintf("failed to execute statement: %s", r[0].Error))
 	}
 }
 
@@ -2025,13 +1139,78 @@ func asJSON(v interface{}) string {
 	return string(b)
 }
 
+// mustTempPath returns a path which can be used for a temporary file or directory.
+// No file will exist at the path.
+func mustTempPath() string {
+	tmpfile, err := os.CreateTemp("", "rqlite-db-test")
+	if err != nil {
+		panic(err.Error())
+	}
+	tmpfile.Close()
+	if err := os.Remove(tmpfile.Name()); err != nil {
+		panic(err.Error())
+	}
+	return tmpfile.Name()
+}
+
 // mustTempFile returns a path to a temporary file in directory dir. It is up to the
 // caller to remove the file once it is no longer needed.
 func mustTempFile() string {
-	tmpfile, err := ioutil.TempFile("", "rqlite-db-test")
+	tmpfile, err := os.CreateTemp("", "rqlite-db-test")
 	if err != nil {
 		panic(err.Error())
 	}
 	tmpfile.Close()
 	return tmpfile.Name()
+}
+
+func mustTempDir() string {
+	tmpdir, err := os.MkdirTemp("", "rqlite-db-test")
+	if err != nil {
+		panic(err.Error())
+	}
+	return tmpdir
+}
+
+// function which copies a src file to a dst file, panics if any error
+func mustCopyFile(dst, src string) {
+	srcFile, err := os.Open(src)
+	if err != nil {
+		panic(err)
+	}
+	defer srcFile.Close()
+
+	dstFile, err := os.Create(dst)
+	if err != nil {
+		panic(err)
+	}
+	defer dstFile.Close()
+
+	_, err = io.Copy(dstFile, srcFile)
+	if err != nil {
+		panic(err)
+	}
+}
+
+func mustCreateClosedFile(path string) {
+	f, err := os.Create(path)
+	if err != nil {
+		panic("failed to create file")
+	}
+	if err := f.Close(); err != nil {
+		panic("failed to close file")
+	}
+}
+
+func mustStat(path string) os.FileInfo {
+	fi, err := os.Stat(path)
+	if err != nil {
+		panic("failed to stat file")
+	}
+	return fi
+}
+
+func mustFileSize(path string) int64 {
+	fi := mustStat(path)
+	return fi.Size()
 }
