@@ -5,9 +5,7 @@ package auth
 import (
 	"encoding/json"
 	"io"
-	"sync"
-
-	"golang.org/x/crypto/bcrypt"
+	"os"
 )
 
 const (
@@ -42,43 +40,6 @@ type BasicAuther interface {
 	BasicAuth() (string, string, bool)
 }
 
-// HashCache store hash values for users. Safe for use from multiple goroutines.
-type HashCache struct {
-	mu sync.RWMutex
-	m  map[string]map[string]struct{}
-}
-
-// NewHashCache returns a instantiated HashCache
-func NewHashCache() *HashCache {
-	return &HashCache{
-		m: make(map[string]map[string]struct{}),
-	}
-}
-
-// Check returns whether hash is valid for username.
-func (h *HashCache) Check(username, hash string) bool {
-	h.mu.RLock()
-	defer h.mu.RUnlock()
-	m, ok := h.m[username]
-	if !ok {
-		return false
-	}
-
-	_, ok = m[hash]
-	return ok
-}
-
-// Store stores the given hash as a valid hash for username.
-func (h *HashCache) Store(username, hash string) {
-	h.mu.Lock()
-	defer h.mu.Unlock()
-	_, ok := h.m[username]
-	if !ok {
-		h.m[username] = make(map[string]struct{})
-	}
-	h.m[username][hash] = struct{}{}
-}
-
 // Credential represents authentication and authorization configuration for a single user.
 type Credential struct {
 	Username string   `json:"username,omitempty"`
@@ -90,19 +51,26 @@ type Credential struct {
 type CredentialsStore struct {
 	store map[string]string
 	perms map[string]map[string]bool
-
-	UseCache  bool
-	hashCache *HashCache
 }
 
 // NewCredentialsStore returns a new instance of a CredentialStore.
 func NewCredentialsStore() *CredentialsStore {
 	return &CredentialsStore{
-		store:     make(map[string]string),
-		perms:     make(map[string]map[string]bool),
-		hashCache: NewHashCache(),
-		UseCache:  true,
+		store: make(map[string]string),
+		perms: make(map[string]map[string]bool),
 	}
+}
+
+// NewCredentialsStoreFromFile returns a new instance of a CredentialStore loaded from a file.
+func NewCredentialsStoreFromFile(path string) (*CredentialsStore, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return nil, err
+	}
+	defer f.Close()
+
+	c := NewCredentialsStore()
+	return c, c.Load(f)
 }
 
 // Load loads credential information from a reader.
@@ -139,31 +107,7 @@ func (c *CredentialsStore) Load(r io.Reader) error {
 // Check returns true if the password is correct for the given username.
 func (c *CredentialsStore) Check(username, password string) bool {
 	pw, ok := c.store[username]
-	if !ok {
-		return false
-	}
-
-	// Simple match with plaintext password in creds?
-	if password == pw {
-		return true
-	}
-
-	// Maybe the given password is a hash -- check if the hash is good
-	// for the given user. We use a cache to avoid recomputing a value we
-	// previously computed (at substantial compute cost).
-	if c.UseCache && c.hashCache.Check(username, password) {
-		return true
-	}
-
-	// Next, what's in the file may be hashed, so hash the given password
-	// and compare.
-	if bcrypt.CompareHashAndPassword([]byte(pw), []byte(password)) != nil {
-		return false
-	}
-
-	// It's good -- cache that result for this user.
-	c.hashCache.Store(username, password)
-	return true
+	return ok && pw == password
 }
 
 // Password returns the password for the given user.
@@ -227,12 +171,12 @@ func (c *CredentialsStore) AA(username, password, perm string) bool {
 		return true
 	}
 
-	// At this point a username needs to have been supplied
+	// At this point a username needs to have been supplied.
 	if username == "" {
 		return false
 	}
 
-	// Are the creds good?
+	// Authenticate the user.
 	if !c.Check(username, password) {
 		return false
 	}
@@ -246,8 +190,5 @@ func (c *CredentialsStore) AA(username, password, perm string) bool {
 // in the request, it returns false.
 func (c *CredentialsStore) HasPermRequest(b BasicAuther, perm string) bool {
 	username, _, ok := b.BasicAuth()
-	if !ok {
-		return false
-	}
-	return c.HasPerm(username, perm)
+	return ok && c.HasPerm(username, perm)
 }
