@@ -2,6 +2,7 @@ package cluster
 
 import (
 	"bytes"
+	"compress/gzip"
 	"encoding/binary"
 	"errors"
 	"fmt"
@@ -10,15 +11,16 @@ import (
 	"testing"
 	"time"
 
-	"github.com/rqlite/rqlite/command"
-	"github.com/rqlite/rqlite/command/encoding"
+	"github.com/rqlite/rqlite/v8/cluster/proto"
+	"github.com/rqlite/rqlite/v8/command/encoding"
+	command "github.com/rqlite/rqlite/v8/command/proto"
 )
 
 const shortWait = 1 * time.Second
 const longWait = 5 * time.Second
 
 var (
-	NO_CREDS = (*Credentials)(nil)
+	NO_CREDS = (*proto.Credentials)(nil)
 )
 
 func Test_ServiceExecute(t *testing.T) {
@@ -286,7 +288,7 @@ func Test_ServiceBackup(t *testing.T) {
 		if br.Format != command.BackupRequest_BACKUP_REQUEST_FORMAT_BINARY {
 			t.Fatalf("wrong backup format requested")
 		}
-		dst.Write(testData)
+		dst.Write(mustGZIPCompress(testData))
 		return nil
 	}
 
@@ -402,6 +404,128 @@ func Test_ServiceRemoveNode(t *testing.T) {
 	}
 }
 
+func Test_ServiceJoinNode(t *testing.T) {
+	ln, mux := mustNewMux()
+	go mux.Serve()
+	tn := mux.Listen(1) // Could be any byte value.
+	db := mustNewMockDatabase()
+	mgr := mustNewMockManager()
+	cred := mustNewMockCredentialStore()
+	s := New(tn, db, mgr, cred)
+	if s == nil {
+		t.Fatalf("failed to create cluster service")
+	}
+
+	c := NewClient(mustNewDialer(1, false, false), 30*time.Second)
+
+	if err := s.Open(); err != nil {
+		t.Fatalf("failed to open cluster service: %s", err.Error())
+	}
+
+	expNodeAddr := "test-node-addr"
+	called := false
+	mgr.joinFn = func(jr *command.JoinRequest) error {
+		called = true
+		if jr.Address != expNodeAddr {
+			t.Fatalf("node address is wrong, exp: %s, got %s", expNodeAddr, jr.Address)
+		}
+		return nil
+	}
+
+	req := &command.JoinRequest{
+		Address: expNodeAddr,
+	}
+	err := c.Join(req, s.Addr(), nil, longWait)
+	if err != nil {
+		t.Fatalf("failed to join node: %s", err.Error())
+	}
+
+	if !called {
+		t.Fatal("JoinNode not called on manager")
+	}
+
+	// Clean up resources
+	if err := ln.Close(); err != nil {
+		t.Fatalf("failed to close Mux's listener: %s", err)
+	}
+	if err := s.Close(); err != nil {
+		t.Fatalf("failed to close cluster service")
+	}
+}
+
+// Test_ServiceJoinNodeForwarded ensures that a JoinNode request is forwarded
+// to the leader if the node receiving the request is not the leader.
+func Test_ServiceJoinNodeForwarded(t *testing.T) {
+	headerByte := byte(1)
+	cred := mustNewMockCredentialStore()
+	c := NewClient(mustNewDialer(headerByte, false, false), 30*time.Second)
+	leaderJoinCalled := false
+
+	// Create the Leader service.
+	lnL, muxL := mustNewMux()
+	go muxL.Serve()
+	tnL := muxL.Listen(headerByte)
+	dbL := mustNewMockDatabase()
+	mgrL := mustNewMockManager()
+	sL := New(tnL, dbL, mgrL, cred)
+	if sL == nil {
+		t.Fatalf("failed to create cluster service for Leader")
+	}
+	mgrL.joinFn = func(jr *command.JoinRequest) error {
+		leaderJoinCalled = true
+		return nil
+	}
+	if err := sL.Open(); err != nil {
+		t.Fatalf("failed to open cluster service on Leader: %s", err.Error())
+	}
+
+	// Create the Follower service.
+	lnF, muxF := mustNewMux()
+	go muxF.Serve()
+	tnF := muxF.Listen(headerByte)
+	dbF := mustNewMockDatabase()
+	mgrF := mustNewMockManager()
+	sF := New(tnF, dbF, mgrF, cred)
+	if sL == nil {
+		t.Fatalf("failed to create cluster service for Follower")
+	}
+	mgrF.joinFn = func(jr *command.JoinRequest) error {
+		return fmt.Errorf("not leader")
+	}
+	mgrF.leaderAddrFn = func() (string, error) {
+		return sL.Addr(), nil
+	}
+	if err := sF.Open(); err != nil {
+		t.Fatalf("failed to open cluster service on Follower: %s", err.Error())
+	}
+
+	req := &command.JoinRequest{
+		Address: "some client",
+	}
+	err := c.Join(req, sF.Addr(), nil, longWait)
+	if err != nil {
+		t.Fatalf("failed to join node: %s", err.Error())
+	}
+
+	if !leaderJoinCalled {
+		t.Fatal("JoinNode not called on leader")
+	}
+
+	// Clean up resources
+	if err := lnL.Close(); err != nil {
+		t.Fatalf("failed to close Mux's listener: %s", err)
+	}
+	if err := sL.Close(); err != nil {
+		t.Fatalf("failed to close cluster service")
+	}
+	if err := lnF.Close(); err != nil {
+		t.Fatalf("failed to close Mux's listener: %s", err)
+	}
+	if err := sF.Close(); err != nil {
+		t.Fatalf("failed to close cluster service")
+	}
+}
+
 // Test_BinaryEncoding_Backwards ensures that software earlier than v6.6.2
 // can communicate with v6.6.2+ releases. v6.6.2 increased the maximum size
 // of cluster responses.
@@ -458,6 +582,24 @@ func queryRequestFromStrings(s []string) *command.QueryRequest {
 	}
 }
 
+func executeQueryRequestFromString(s string) *command.ExecuteQueryRequest {
+	return executeQueryRequestFromStrings([]string{s})
+}
+
+func executeQueryRequestFromStrings(s []string) *command.ExecuteQueryRequest {
+	stmts := make([]*command.Statement, len(s))
+	for i := range s {
+		stmts[i] = &command.Statement{
+			Sql: s[i],
+		}
+	}
+	return &command.ExecuteQueryRequest{
+		Request: &command.Request{
+			Statements: stmts,
+		},
+	}
+}
+
 func backupRequestBinary(leader bool) *command.BackupRequest {
 	return &command.BackupRequest{
 		Format: command.BackupRequest_BACKUP_REQUEST_FORMAT_BINARY,
@@ -475,6 +617,18 @@ func removeNodeRequest(id string) *command.RemoveNodeRequest {
 	return &command.RemoveNodeRequest{
 		Id: id,
 	}
+}
+
+func mustGZIPCompress(b []byte) []byte {
+	var buf bytes.Buffer
+	zw := gzip.NewWriter(&buf)
+	if _, err := zw.Write(b); err != nil {
+		panic(fmt.Sprintf("failed to compress data: %s", err.Error()))
+	}
+	if err := zw.Close(); err != nil {
+		panic(fmt.Sprintf("failed to close gzip writer: %s", err.Error()))
+	}
+	return buf.Bytes()
 }
 
 func asJSON(v interface{}) string {
