@@ -8,9 +8,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/rqlite/rqlite/rtls"
-	"github.com/rqlite/rqlite/tcp"
-	"github.com/rqlite/rqlite/testdata/x509"
+	"github.com/rqlite/rqlite/v8/rtls"
+	"github.com/rqlite/rqlite/v8/tcp"
+	"github.com/rqlite/rqlite/v8/testdata/x509"
 )
 
 func Test_NewServiceSetGetNodeAPIAddrMuxed(t *testing.T) {
@@ -101,9 +101,9 @@ func mustNewTLSMux() (net.Listener, *tcp.Mux) {
 		panic("failed to create mock listener")
 	}
 
-	cert := x509.CertFile("")
+	cert := x509.CertExampleDotComFile("")
 	defer os.Remove(cert)
-	key := x509.KeyFile("")
+	key := x509.KeyExampleDotComFile("")
 	defer os.Remove(key)
 
 	mux, err := tcp.NewTLSMux(ln, nil, cert, key, "", true, false)
@@ -118,7 +118,7 @@ func mustNewDialer(header byte, remoteEncrypted, skipVerify bool) *tcp.Dialer {
 	var tlsConfig *tls.Config
 	var err error
 	if remoteEncrypted {
-		tlsConfig, err = rtls.CreateClientConfig("", "", "", skipVerify, false)
+		tlsConfig, err = rtls.CreateClientConfig("", "", rtls.NoCACert, rtls.NoServerName, skipVerify)
 		if err != nil {
 			panic(fmt.Sprintf("failed to create client TLS config: %s", err))
 		}
