@@ -6,11 +6,13 @@ import (
 	"io"
 	"net"
 	"os"
+	"sync"
 	"testing"
 	"time"
 
-	"github.com/rqlite/rqlite/command"
-	"github.com/rqlite/rqlite/testdata/x509"
+	"github.com/rqlite/rqlite/v8/cluster/proto"
+	command "github.com/rqlite/rqlite/v8/command/proto"
+	"github.com/rqlite/rqlite/v8/testdata/x509"
 )
 
 func Test_NewServiceOpenClose(t *testing.T) {
@@ -258,6 +260,126 @@ func Test_NewServiceTestExecuteQueryAuth(t *testing.T) {
 	}
 }
 
+func Test_NewServiceNotify(t *testing.T) {
+	ml := mustNewMockTransport()
+	mm := mustNewMockManager()
+	var wg sync.WaitGroup
+
+	wg.Add(1)
+	mm.notifyFn = func(n *command.NotifyRequest) error {
+		defer wg.Done()
+		if n.Id != "foo" {
+			t.Fatalf("failed to get correct node ID, exp %s, got %s", "foo", n.Id)
+		}
+		if n.Address != "localhost" {
+			t.Fatalf("failed to get correct node address, exp %s, got %s", "localhost", n.Address)
+		}
+		return nil
+	}
+
+	credStr := mustNewMockCredentialStore()
+	s := New(ml, mustNewMockDatabase(), mm, credStr)
+	if s == nil {
+		t.Fatalf("failed to create cluster service")
+	}
+
+	if err := s.Open(); err != nil {
+		t.Fatalf("failed to open cluster service")
+	}
+
+	// Create a notify request.
+	nr := &command.NotifyRequest{
+		Id:      "foo",
+		Address: "localhost",
+	}
+
+	// Test by connecting to itself.
+	c := NewClient(ml, 30*time.Second)
+	err := c.Notify(nr, s.Addr(), nil, 5*time.Second)
+	if err != nil {
+		t.Fatalf("failed to notify node: %s", err)
+	}
+
+	// Ensure that the notify function was called.
+	wg.Wait()
+
+	// Test when auth is enabled
+	credStr.HasPermOK = false
+	err = c.Notify(nr, s.Addr(), nil, 5*time.Second)
+	if err == nil {
+		t.Fatal("should have failed to notify node due to lack of auth")
+	}
+	if err.Error() != "unauthorized" {
+		t.Fatalf("failed to get correct error, exp %s, got %s", "unauthorized", err.Error())
+	}
+
+	if err := s.Close(); err != nil {
+		t.Fatalf("failed to close cluster service")
+	}
+}
+
+func Test_NewServiceJoin(t *testing.T) {
+	ml := mustNewMockTransport()
+	mm := mustNewMockManager()
+	var wg sync.WaitGroup
+
+	wg.Add(1)
+	mm.joinFn = func(j *command.JoinRequest) error {
+		defer wg.Done()
+		if j.Id != "foo" {
+			t.Fatalf("failed to get correct node ID, exp %s, got %s", "foo", j.Id)
+		}
+		if j.Address != "localhost" {
+			t.Fatalf("failed to get correct node address, exp %s, got %s", "localhost", j.Address)
+		}
+		if !j.Voter {
+			t.Fatalf("failed to get correct voter setting, exp %t, got %t", true, j.Voter)
+		}
+		return nil
+	}
+
+	credStr := mustNewMockCredentialStore()
+	s := New(ml, mustNewMockDatabase(), mm, credStr)
+	if s == nil {
+		t.Fatalf("failed to create cluster service")
+	}
+
+	if err := s.Open(); err != nil {
+		t.Fatalf("failed to open cluster service")
+	}
+
+	// Create a Join request.
+	jr := &command.JoinRequest{
+		Id:      "foo",
+		Address: "localhost",
+		Voter:   true,
+	}
+
+	// Test by connecting to itself.
+	c := NewClient(ml, 30*time.Second)
+	err := c.Join(jr, s.Addr(), nil, 5*time.Second)
+	if err != nil {
+		t.Fatalf("failed to join node: %s", err)
+	}
+
+	// Ensure the join function was called.
+	wg.Wait()
+
+	// Test when auth is enabled
+	credStr.HasPermOK = false
+	err = c.Join(jr, s.Addr(), nil, 5*time.Second)
+	if err == nil {
+		t.Fatal("should have failed to join node due to lack of auth")
+	}
+	if err.Error() != "unauthorized" {
+		t.Fatalf("failed to get correct error, exp %s, got %s", "unauthorized", err.Error())
+	}
+
+	if err := s.Close(); err != nil {
+		t.Fatalf("failed to close cluster service")
+	}
+}
+
 type mockTransport struct {
 	tn              net.Listener
 	remoteEncrypted bool
@@ -313,6 +435,7 @@ func mustNewMockTLSTransport() *mockTransport {
 type mockDatabase struct {
 	executeFn func(er *command.ExecuteRequest) ([]*command.ExecuteResult, error)
 	queryFn   func(qr *command.QueryRequest) ([]*command.QueryRows, error)
+	requestFn func(rr *command.ExecuteQueryRequest) ([]*command.ExecuteQueryResponse, error)
 	backupFn  func(br *command.BackupRequest, dst io.Writer) error
 	loadFn    func(lr *command.LoadRequest) error
 }
@@ -323,6 +446,13 @@ func (m *mockDatabase) Execute(er *command.ExecuteRequest) ([]*command.ExecuteRe
 
 func (m *mockDatabase) Query(qr *command.QueryRequest) ([]*command.QueryRows, error) {
 	return m.queryFn(qr)
+}
+
+func (m *mockDatabase) Request(rr *command.ExecuteQueryRequest) ([]*command.ExecuteQueryResponse, error) {
+	if m.requestFn == nil {
+		return []*command.ExecuteQueryResponse{}, nil
+	}
+	return m.requestFn(rr)
 }
 
 func (m *mockDatabase) Backup(br *command.BackupRequest, dst io.Writer) error {
@@ -351,6 +481,9 @@ func mustNewMockDatabase() *mockDatabase {
 
 type MockManager struct {
 	removeNodeFn func(rn *command.RemoveNodeRequest) error
+	notifyFn     func(n *command.NotifyRequest) error
+	joinFn       func(j *command.JoinRequest) error
+	leaderAddrFn func() (string, error)
 }
 
 func (m *MockManager) Remove(rn *command.RemoveNodeRequest) error {
@@ -360,6 +493,27 @@ func (m *MockManager) Remove(rn *command.RemoveNodeRequest) error {
 	return m.removeNodeFn(rn)
 }
 
+func (m *MockManager) Notify(n *command.NotifyRequest) error {
+	if m.notifyFn == nil {
+		return nil
+	}
+	return m.notifyFn(n)
+}
+
+func (m *MockManager) Join(j *command.JoinRequest) error {
+	if m.joinFn == nil {
+		return nil
+	}
+	return m.joinFn(j)
+}
+
+func (m *MockManager) LeaderAddr() (string, error) {
+	if m.leaderAddrFn == nil {
+		return "", nil
+	}
+	return m.leaderAddrFn()
+}
+
 func mustNewMockManager() *MockManager {
 	return &MockManager{}
 }
@@ -367,9 +521,9 @@ func mustNewMockManager() *MockManager {
 func mustCreateTLSConfig() *tls.Config {
 	var err error
 
-	certFile := x509.CertFile("")
+	certFile := x509.CertExampleDotComFile("")
 	defer os.Remove(certFile)
-	keyFile := x509.KeyFile("")
+	keyFile := x509.KeyExampleDotComFile("")
 	defer os.Remove(keyFile)
 
 	config := &tls.Config{
@@ -404,8 +558,8 @@ func mustNewMockCredentialStore() *mockCredentialStore {
 	return &mockCredentialStore{HasPermOK: true}
 }
 
-func makeCredentials(username, password string) *Credentials {
-	return &Credentials{
+func makeCredentials(username, password string) *proto.Credentials {
+	return &proto.Credentials{
 		Username: username,
 		Password: password,
 	}
