@@ -2,13 +2,13 @@
 package main
 
 import (
+	"context"
 	"crypto/tls"
 	"fmt"
 	"log"
 	"net"
 	"os"
 	"os/signal"
-	"path/filepath"
 	"runtime"
 	"strings"
 	"syscall"
@@ -18,15 +18,18 @@ import (
 	"github.com/rqlite/rqlite-disco-clients/dns"
 	"github.com/rqlite/rqlite-disco-clients/dnssrv"
 	etcd "github.com/rqlite/rqlite-disco-clients/etcd"
-	"github.com/rqlite/rqlite/auth"
-	"github.com/rqlite/rqlite/cluster"
-	"github.com/rqlite/rqlite/cmd"
-	"github.com/rqlite/rqlite/db"
-	"github.com/rqlite/rqlite/disco"
-	httpd "github.com/rqlite/rqlite/http"
-	"github.com/rqlite/rqlite/rtls"
-	"github.com/rqlite/rqlite/store"
-	"github.com/rqlite/rqlite/tcp"
+	"github.com/rqlite/rqlite/v8/auth"
+	"github.com/rqlite/rqlite/v8/auto/backup"
+	"github.com/rqlite/rqlite/v8/auto/restore"
+	"github.com/rqlite/rqlite/v8/aws"
+	"github.com/rqlite/rqlite/v8/cluster"
+	"github.com/rqlite/rqlite/v8/cmd"
+	"github.com/rqlite/rqlite/v8/db"
+	"github.com/rqlite/rqlite/v8/disco"
+	httpd "github.com/rqlite/rqlite/v8/http"
+	"github.com/rqlite/rqlite/v8/rtls"
+	"github.com/rqlite/rqlite/v8/store"
+	"github.com/rqlite/rqlite/v8/tcp"
 )
 
 const logo = `
@@ -43,7 +46,9 @@ const logo = `
 
 const name = `rqlited`
 const desc = `rqlite is a lightweight, distributed relational database, which uses SQLite as its
-storage engine. It provides an easy-to-use, fault-tolerant store for relational data.`
+storage engine. It provides an easy-to-use, fault-tolerant store for relational data.
+
+Visit https://www.rqlite.io to learn more.`
 
 func init() {
 	log.SetFlags(log.LstdFlags)
@@ -61,9 +66,10 @@ func main() {
 	if err != nil {
 		log.Fatalf("failed to parse command-line flags: %s", err.Error())
 	}
-
-	// Display logo.
 	fmt.Print(logo)
+
+	mainCtx, mainCancel := context.WithCancel(context.Background())
+	defer mainCancel()
 
 	// Configure logging and pump out initial message.
 	log.Printf("%s starting, version %s, SQLite %s, commit %s, branch %s, compiler %s", name, cmd.Version,
@@ -84,13 +90,51 @@ func main() {
 	if err != nil {
 		log.Fatalf("failed to start node mux: %s", err.Error())
 	}
-	raftTn := mux.Listen(cluster.MuxRaftHeader)
+
+	// Raft internode layer
+	raftLn := mux.Listen(cluster.MuxRaftHeader)
 	log.Printf("Raft TCP mux Listener registered with byte header %d", cluster.MuxRaftHeader)
+	raftDialer, err := cluster.CreateRaftDialer(cfg.NodeX509Cert, cfg.NodeX509Key, cfg.NodeX509CACert,
+		cfg.NodeVerifyServerName, cfg.NoNodeVerify)
+	if err != nil {
+		log.Fatalf("failed to create Raft dialer: %s", err.Error())
+	}
+	raftTn := tcp.NewLayer(raftLn, raftDialer)
 
 	// Create the store.
 	str, err := createStore(cfg, raftTn)
 	if err != nil {
 		log.Fatalf("failed to create store: %s", err.Error())
+	}
+
+	// Install the auto-restore data, if necessary.
+	if cfg.AutoRestoreFile != "" {
+		hd, err := store.HasData(str.Path())
+		if err != nil {
+			log.Fatalf("failed to check for existing data: %s", err.Error())
+		}
+		if hd {
+			log.Printf("auto-restore requested, but data already exists in %s, skipping", str.Path())
+		} else {
+			log.Printf("auto-restore requested, initiating download")
+			start := time.Now()
+			path, errOK, err := restore.DownloadFile(mainCtx, cfg.AutoRestoreFile)
+			if err != nil {
+				var b strings.Builder
+				b.WriteString(fmt.Sprintf("failed to download auto-restore file: %s", err.Error()))
+				if errOK {
+					b.WriteString(", continuing with node startup anyway")
+					log.Print(b.String())
+				} else {
+					log.Fatal(b.String())
+				}
+			} else {
+				log.Printf("auto-restore file downloaded in %s", time.Since(start))
+				if err := str.SetRestorePath(path); err != nil {
+					log.Fatalf("failed to preload auto-restore data: %s", err.Error())
+				}
+			}
+		}
 	}
 
 	// Get any credential store.
@@ -128,13 +172,14 @@ func main() {
 
 	// Register remaining status providers.
 	httpServ.RegisterStatus("cluster", clstrServ)
+	httpServ.RegisterStatus("network", tcp.NetworkReporter{})
 
 	// Create the cluster!
 	nodes, err := str.Nodes()
 	if err != nil {
 		log.Fatalf("failed to get nodes %s", err.Error())
 	}
-	if err := createCluster(cfg, len(nodes) > 0, str, httpServ, credStr); err != nil {
+	if err := createCluster(cfg, len(nodes) > 0, clstrClient, str, httpServ, credStr); err != nil {
 		log.Fatalf("clustering failure: %s", err.Error())
 	}
 
@@ -143,10 +188,36 @@ func main() {
 	h, p, _ := net.SplitHostPort(cfg.HTTPAdv)
 	log.Printf("connect using the command-line tool via 'rqlite -H %s -p %s'", h, p)
 
+	// Start any requested auto-backups
+	backupSrvStx, backupSrvCancel := context.WithCancel(mainCtx)
+	backupSrv, err := startAutoBackups(backupSrvStx, cfg, str)
+	if err != nil {
+		log.Fatalf("failed to start auto-backups: %s", err.Error())
+	}
+	if backupSrv != nil {
+		httpServ.RegisterStatus("auto_backups", backupSrv)
+	}
+
 	// Block until signalled.
 	terminate := make(chan os.Signal, 1)
 	signal.Notify(terminate, syscall.SIGINT, syscall.SIGTERM, os.Interrupt)
-	<-terminate
+	sig := <-terminate
+	log.Printf(`received signal "%s", shutting down`, sig.String())
+
+	// Stop the HTTP server first, so clients get notification as soon as
+	// possible that the node is going away.
+	httpServ.Close()
+
+	if cfg.RaftClusterRemoveOnShutdown {
+		remover := cluster.NewRemover(clstrClient, 5*time.Second, str)
+		remover.SetCredentials(cluster.CredentialsFor(credStr, cfg.JoinAs))
+		log.Printf("initiating removal of this node from cluster before shutdown")
+		if err := remover.Do(cfg.NodeID, true); err != nil {
+			log.Fatalf("failed to remove this node from cluster before shutdown: %s", err.Error())
+		} else {
+			log.Printf("removed this node successfully from cluster before shutdown")
+		}
+	}
 
 	if cfg.RaftStepdownOnShutdown {
 		if str.IsLeader() {
@@ -157,7 +228,7 @@ func main() {
 		str.Stepdown(true)
 	}
 
-	httpServ.Close()
+	backupSrvCancel()
 	if err := str.Close(true); err != nil {
 		log.Printf("failed to close store: %s", err.Error())
 	}
@@ -167,12 +238,30 @@ func main() {
 	log.Println("rqlite server stopped")
 }
 
-func createStore(cfg *Config, ln *tcp.Layer) (*store.Store, error) {
-	dataPath, err := filepath.Abs(cfg.DataPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to determine absolute data path: %s", err.Error())
+func startAutoBackups(ctx context.Context, cfg *Config, str *store.Store) (*backup.Uploader, error) {
+	if cfg.AutoBackupFile == "" {
+		return nil, nil
 	}
-	dbConf := store.NewDBConfig(!cfg.OnDisk)
+
+	b, err := backup.ReadConfigFile(cfg.AutoBackupFile)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read auto-backup file: %s", err.Error())
+	}
+
+	uCfg, s3cfg, err := backup.Unmarshal(b)
+	if err != nil {
+		return nil, fmt.Errorf("failed to parse auto-backup file: %s", err.Error())
+	}
+	provider := store.NewProvider(str, uCfg.Vacuum, !uCfg.NoCompress)
+	sc := aws.NewS3Client(s3cfg.Endpoint, s3cfg.Region, s3cfg.AccessKeyID, s3cfg.SecretAccessKey,
+		s3cfg.Bucket, s3cfg.Path, s3cfg.ForcePathStyle)
+	u := backup.NewUploader(sc, provider, time.Duration(uCfg.Interval), !uCfg.NoCompress)
+	u.Start(ctx, str.IsLeader)
+	return u, nil
+}
+
+func createStore(cfg *Config, ln *tcp.Layer) (*store.Store, error) {
+	dbConf := store.NewDBConfig()
 	dbConf.OnDiskPath = cfg.OnDiskPath
 	dbConf.FKConstraints = cfg.FKConstraints
 
@@ -183,12 +272,10 @@ func createStore(cfg *Config, ln *tcp.Layer) (*store.Store, error) {
 	})
 
 	// Set optional parameters on store.
-	str.StartupOnDisk = cfg.OnDiskStartup
-	str.SetRequestCompression(cfg.CompressionBatch, cfg.CompressionSize)
 	str.RaftLogLevel = cfg.RaftLogLevel
-	str.NoFreeListSync = cfg.RaftNoFreelistSync
 	str.ShutdownOnRemove = cfg.RaftShutdownOnRemove
 	str.SnapshotThreshold = cfg.RaftSnapThreshold
+	str.SnapshotThresholdWALSize = cfg.RaftSnapThresholdWALSize
 	str.SnapshotInterval = cfg.RaftSnapInterval
 	str.LeaderLeaseTimeout = cfg.RaftLeaderLeaseTimeout
 	str.HeartbeatTimeout = cfg.RaftHeartbeatTimeout
@@ -198,11 +285,10 @@ func createStore(cfg *Config, ln *tcp.Layer) (*store.Store, error) {
 	str.ReapTimeout = cfg.RaftReapNodeTimeout
 	str.ReapReadOnlyTimeout = cfg.RaftReapReadOnlyNodeTimeout
 
-	isNew := store.IsNewNode(dataPath)
-	if isNew {
-		log.Printf("no preexisting node state detected in %s, node may be bootstrapping", dataPath)
+	if store.IsNewNode(cfg.DataPath) {
+		log.Printf("no preexisting node state detected in %s, node may be bootstrapping", cfg.DataPath)
 	} else {
-		log.Printf("preexisting node state detected in %s", dataPath)
+		log.Printf("preexisting node state detected in %s", cfg.DataPath)
 	}
 
 	return str, nil
@@ -244,7 +330,7 @@ func createDiscoService(cfg *Config, str *store.Store) (*disco.Service, error) {
 		return nil, fmt.Errorf("invalid disco service: %s", cfg.DiscoMode)
 	}
 
-	return disco.NewService(c, str), nil
+	return disco.NewService(c, str, disco.VoterSuffrage(!cfg.RaftNonVoter)), nil
 }
 
 func startHTTPService(cfg *Config, str *store.Store, cltr *cluster.Client, credStr *auth.CredentialsStore) (*httpd.Service, error) {
@@ -254,14 +340,12 @@ func startHTTPService(cfg *Config, str *store.Store, cltr *cluster.Client, credS
 	s.CACertFile = cfg.HTTPx509CACert
 	s.CertFile = cfg.HTTPx509Cert
 	s.KeyFile = cfg.HTTPx509Key
-	s.TLS1011 = cfg.TLS1011
 	s.ClientVerify = cfg.HTTPVerifyClient
-	s.Expvar = cfg.Expvar
-	s.Pprof = cfg.PprofEnabled
 	s.DefaultQueueCap = cfg.WriteQueueCap
 	s.DefaultQueueBatchSz = cfg.WriteQueueBatchSz
 	s.DefaultQueueTimeout = cfg.WriteQueueTimeout
 	s.DefaultQueueTx = cfg.WriteQueueTx
+	s.AllowOrigin = cfg.HTTPAllowOrigin
 	s.BuildInfo = map[string]interface{}{
 		"commit":     cmd.Commit,
 		"branch":     cmd.Branch,
@@ -289,9 +373,9 @@ func startNodeMux(cfg *Config, ln net.Listener) (*tcp.Mux, error) {
 			b.WriteString(fmt.Sprintf(", CA cert %s", cfg.NodeX509CACert))
 		}
 		if cfg.NodeVerifyClient {
-			b.WriteString(", mutual TLS disabled")
-		} else {
 			b.WriteString(", mutual TLS enabled")
+		} else {
+			b.WriteString(", mutual TLS disabled")
 		}
 		log.Println(b.String())
 		mux, err = tcp.NewTLSMux(ln, adv, cfg.NodeX509Cert, cfg.NodeX509Key, cfg.NodeX509CACert,
@@ -311,24 +395,13 @@ func credentialStore(cfg *Config) (*auth.CredentialsStore, error) {
 	if cfg.AuthFile == "" {
 		return nil, nil
 	}
-
-	f, err := os.Open(cfg.AuthFile)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open authentication file %s: %s", cfg.AuthFile, err.Error())
-	}
-
-	cs := auth.NewCredentialsStore()
-	if cs.Load(f); err != nil {
-		return nil, err
-	}
-	return cs, nil
+	return auth.NewCredentialsStoreFromFile(cfg.AuthFile)
 }
 
-func clusterService(cfg *Config, tn cluster.Transport, db cluster.Database, mgr cluster.Manager, credStr *auth.CredentialsStore) (*cluster.Service, error) {
-	c := cluster.New(tn, db, mgr, credStr)
+func clusterService(cfg *Config, ln net.Listener, db cluster.Database, mgr cluster.Manager, credStr *auth.CredentialsStore) (*cluster.Service, error) {
+	c := cluster.New(ln, db, mgr, credStr)
 	c.SetAPIAddr(cfg.HTTPAdv)
 	c.EnableHTTPS(cfg.HTTPx509Cert != "" && cfg.HTTPx509Key != "") // Conditions met for an HTTPS API
-
 	if err := c.Open(); err != nil {
 		return nil, err
 	}
@@ -340,32 +413,24 @@ func createClusterClient(cfg *Config, clstr *cluster.Service) (*cluster.Client, 
 	var err error
 	if cfg.NodeX509Cert != "" || cfg.NodeX509CACert != "" {
 		dialerTLSConfig, err = rtls.CreateClientConfig(cfg.NodeX509Cert, cfg.NodeX509Key,
-			cfg.NodeX509CACert, cfg.NoNodeVerify, cfg.TLS1011)
+			cfg.NodeX509CACert, cfg.NodeVerifyServerName, cfg.NoNodeVerify)
 		if err != nil {
-			log.Fatalf("failed to create TLS config for cluster dialer: %s", err.Error())
+			return nil, fmt.Errorf("failed to create TLS config for cluster dialer: %s", err.Error())
 		}
 	}
 	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, dialerTLSConfig)
 	clstrClient := cluster.NewClient(clstrDialer, cfg.ClusterConnectTimeout)
 	if err := clstrClient.SetLocal(cfg.RaftAdv, clstr); err != nil {
-		log.Fatalf("failed to set cluster client local parameters: %s", err.Error())
+		return nil, fmt.Errorf("failed to set cluster client local parameters: %s", err.Error())
 	}
 	return clstrClient, nil
 }
 
-func createCluster(cfg *Config, hasPeers bool, str *store.Store,
-	httpServ *httpd.Service, credStr *auth.CredentialsStore) error {
-	var tlsConfig *tls.Config
-	var err error
-	if cfg.HTTPx509Cert != "" || cfg.HTTPx509CACert != "" {
-		tlsConfig, err = rtls.CreateClientConfig(cfg.HTTPx509Cert, cfg.HTTPx509Key, cfg.HTTPx509CACert,
-			cfg.NoHTTPVerify, cfg.TLS1011)
-		if err != nil {
-			return fmt.Errorf("failed to create TLS client config for cluster: %s", err.Error())
-		}
-	}
-
+func createCluster(cfg *Config, hasPeers bool, client *cluster.Client, str *store.Store, httpServ *httpd.Service, credStr *auth.CredentialsStore) error {
 	joins := cfg.JoinAddresses()
+	if err := networkCheckJoinAddrs(joins); err != nil {
+		return err
+	}
 	if joins == nil && cfg.DiscoMode == "" && !hasPeers {
 		if cfg.RaftNonVoter {
 			return fmt.Errorf("cannot create a new non-voting node without joining it to an existing cluster")
@@ -379,25 +444,18 @@ func createCluster(cfg *Config, hasPeers bool, str *store.Store,
 		return nil
 	}
 
-	// Prepare the Joiner
-	joiner := cluster.NewJoiner(cfg.JoinSrcIP, cfg.JoinAttempts, cfg.JoinInterval, tlsConfig)
-	if cfg.JoinAs != "" {
-		pw, ok := credStr.Password(cfg.JoinAs)
-		if !ok {
-			return fmt.Errorf("user %s does not exist in credential store", cfg.JoinAs)
-		}
-		joiner.SetBasicAuth(cfg.JoinAs, pw)
-	}
-
 	// Prepare definition of being part of a cluster.
-	isClustered := func() bool {
+	bootDoneFn := func() bool {
 		leader, _ := str.LeaderAddr()
 		return leader != ""
 	}
+	clusterSuf := cluster.VoterSuffrage(!cfg.RaftNonVoter)
 
+	joiner := cluster.NewJoiner(client, cfg.JoinAttempts, cfg.JoinInterval)
+	joiner.SetCredentials(cluster.CredentialsFor(credStr, cfg.JoinAs))
 	if joins != nil && cfg.BootstrapExpect == 0 {
 		// Explicit join operation requested, so do it.
-		j, err := joiner.Do(joins, str.ID(), cfg.RaftAdv, !cfg.RaftNonVoter)
+		j, err := joiner.Do(joins, str.ID(), cfg.RaftAdv, clusterSuf)
 		if err != nil {
 			return fmt.Errorf("failed to join cluster: %s", err.Error())
 		}
@@ -406,21 +464,10 @@ func createCluster(cfg *Config, hasPeers bool, str *store.Store,
 	}
 
 	if joins != nil && cfg.BootstrapExpect > 0 {
-		if hasPeers {
-			log.Println("preexisting node configuration detected, ignoring bootstrap request")
-			return nil
-		}
-
 		// Bootstrap with explicit join addresses requests.
-		bs := cluster.NewBootstrapper(cluster.NewAddressProviderString(joins), tlsConfig)
-		if cfg.JoinAs != "" {
-			pw, ok := credStr.Password(cfg.JoinAs)
-			if !ok {
-				return fmt.Errorf("user %s does not exist in credential store", cfg.JoinAs)
-			}
-			bs.SetBasicAuth(cfg.JoinAs, pw)
-		}
-		return bs.Boot(str.ID(), cfg.RaftAdv, isClustered, cfg.BootstrapExpectTimeout)
+		bs := cluster.NewBootstrapper(cluster.NewAddressProviderString(joins), client)
+		bs.SetCredentials(cluster.CredentialsFor(credStr, cfg.JoinAs))
+		return bs.Boot(str.ID(), cfg.RaftAdv, clusterSuf, bootDoneFn, cfg.BootstrapExpectTimeout)
 	}
 
 	if cfg.DiscoMode == "" {
@@ -429,13 +476,12 @@ func createCluster(cfg *Config, hasPeers bool, str *store.Store,
 		return nil
 	}
 
+	// DNS-based discovery requested. It's OK to proceed with this even if this node
+	// is already part of a cluster. Re-joining and re-notifying other nodes will be
+	// ignored when the node is already part of the cluster.
 	log.Printf("discovery mode: %s", cfg.DiscoMode)
 	switch cfg.DiscoMode {
 	case DiscoModeDNS, DiscoModeDNSSRV:
-		if hasPeers {
-			log.Printf("preexisting node configuration detected, ignoring %s option", cfg.DiscoMode)
-			return nil
-		}
 		rc := cfg.DiscoConfigReader()
 		defer func() {
 			if rc != nil {
@@ -452,7 +498,7 @@ func createCluster(cfg *Config, hasPeers bool, str *store.Store,
 			if err != nil {
 				return fmt.Errorf("error reading DNS configuration: %s", err.Error())
 			}
-			provider = dns.New(dnsCfg)
+			provider = dns.NewWithPort(dnsCfg, cfg.RaftPort())
 
 		} else {
 			dnssrvCfg, err := dnssrv.NewConfigFromReader(rc)
@@ -462,61 +508,65 @@ func createCluster(cfg *Config, hasPeers bool, str *store.Store,
 			provider = dnssrv.New(dnssrvCfg)
 		}
 
-		bs := cluster.NewBootstrapper(provider, tlsConfig)
-		if cfg.JoinAs != "" {
-			pw, ok := credStr.Password(cfg.JoinAs)
-			if !ok {
-				return fmt.Errorf("user %s does not exist in credential store", cfg.JoinAs)
-			}
-			bs.SetBasicAuth(cfg.JoinAs, pw)
-		}
+		bs := cluster.NewBootstrapper(provider, client)
+		bs.SetCredentials(cluster.CredentialsFor(credStr, cfg.JoinAs))
 		httpServ.RegisterStatus("disco", provider)
-		return bs.Boot(str.ID(), cfg.RaftAdv, isClustered, cfg.BootstrapExpectTimeout)
+		return bs.Boot(str.ID(), cfg.RaftAdv, clusterSuf, bootDoneFn, cfg.BootstrapExpectTimeout)
 
 	case DiscoModeEtcdKV, DiscoModeConsulKV:
 		discoService, err := createDiscoService(cfg, str)
 		if err != nil {
 			return fmt.Errorf("failed to start discovery service: %s", err.Error())
 		}
-
-		if !hasPeers {
-			log.Println("no preexisting nodes, registering with discovery service")
-
-			leader, addr, err := discoService.Register(str.ID(), cfg.HTTPURL(), cfg.RaftAdv)
-			if err != nil {
-				return fmt.Errorf("failed to register with discovery service: %s", err.Error())
-			}
-			if leader {
-				log.Println("node registered as leader using discovery service")
-				if err := str.Bootstrap(store.NewServer(str.ID(), str.Addr(), true)); err != nil {
-					return fmt.Errorf("failed to bootstrap single new node: %s", err.Error())
-				}
-			} else {
-				for {
-					log.Printf("discovery service returned %s as join address", addr)
-					if j, err := joiner.Do([]string{addr}, str.ID(), cfg.RaftAdv, !cfg.RaftNonVoter); err != nil {
-						log.Printf("failed to join cluster at %s: %s", addr, err.Error())
-
-						time.Sleep(time.Second)
-						_, addr, err = discoService.Register(str.ID(), cfg.HTTPURL(), cfg.RaftAdv)
-						if err != nil {
-							log.Printf("failed to get updated leader: %s", err.Error())
-						}
-						continue
-					} else {
-						log.Println("successfully joined cluster at", j)
-						break
-					}
-				}
-			}
-		} else {
-			log.Println("preexisting node configuration detected, not registering with discovery service")
-		}
+		// Safe to start reporting before doing registration. If the node hasn't bootstrapped
+		// yet, or isn't leader, reporting will just be a no-op until something changes.
 		go discoService.StartReporting(cfg.NodeID, cfg.HTTPURL(), cfg.RaftAdv)
 		httpServ.RegisterStatus("disco", discoService)
 
+		if hasPeers {
+			log.Printf("preexisting node configuration detected, not registering with discovery service")
+			return nil
+		}
+		log.Println("no preexisting nodes, registering with discovery service")
+
+		leader, addr, err := discoService.Register(str.ID(), cfg.HTTPURL(), cfg.RaftAdv)
+		if err != nil {
+			return fmt.Errorf("failed to register with discovery service: %s", err.Error())
+		}
+		if leader {
+			log.Println("node registered as leader using discovery service")
+			if err := str.Bootstrap(store.NewServer(str.ID(), str.Addr(), true)); err != nil {
+				return fmt.Errorf("failed to bootstrap single new node: %s", err.Error())
+			}
+		} else {
+			for {
+				log.Printf("discovery service returned %s as join address", addr)
+				if j, err := joiner.Do([]string{addr}, str.ID(), cfg.RaftAdv, clusterSuf); err != nil {
+					log.Printf("failed to join cluster at %s: %s", addr, err.Error())
+
+					time.Sleep(time.Second)
+					_, addr, err = discoService.Register(str.ID(), cfg.HTTPURL(), cfg.RaftAdv)
+					if err != nil {
+						log.Printf("failed to get updated leader: %s", err.Error())
+					}
+					continue
+				} else {
+					log.Println("successfully joined cluster at", j)
+					break
+				}
+			}
+		}
+
 	default:
 		return fmt.Errorf("invalid disco mode %s", cfg.DiscoMode)
+	}
+	return nil
+}
+
+func networkCheckJoinAddrs(joinAddrs []string) error {
+	log.Println("checking that any supplied join addresses don't serve HTTP(S)")
+	if addr, ok := httpd.AnyServingHTTP(joinAddrs); ok {
+		return fmt.Errorf("join address %s appears to be serving HTTP when it should be Raft", addr)
 	}
 	return nil
 }
