@@ -6,7 +6,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/rqlite/rqlite/command"
+	"github.com/rqlite/rqlite/v8/command/proto"
 )
 
 var (
@@ -35,13 +35,55 @@ type Rows struct {
 // AssociativeRows represents the outcome of an operation that returns query data.
 type AssociativeRows struct {
 	Types map[string]string        `json:"types,omitempty"`
-	Rows  []map[string]interface{} `json:"rows,omitempty"`
+	Rows  []map[string]interface{} `json:"rows"`
 	Error string                   `json:"error,omitempty"`
 	Time  float64                  `json:"time,omitempty"`
 }
 
+// ResultWithRows represents the outcome of an operation that changes rows, but also
+// includes an nil rows object, so clients can distinguish between a query and execute
+// result.
+type ResultWithRows struct {
+	Result
+	Rows []map[string]interface{} `json:"rows"`
+}
+
+// NewResultRowsFromExecuteQueryResponse returns an API object from an
+// ExecuteQueryResponse.
+func NewResultRowsFromExecuteQueryResponse(e *proto.ExecuteQueryResponse) (interface{}, error) {
+	if er := e.GetE(); er != nil {
+		return NewResultFromExecuteResult(er)
+	} else if qr := e.GetQ(); qr != nil {
+		return NewRowsFromQueryRows(qr)
+	} else if err := e.GetError(); err != "" {
+		return map[string]string{
+			"error": err,
+		}, nil
+	}
+	return nil, errors.New("no ExecuteResult, QueryRows, or Error")
+}
+
+func NewAssociativeResultRowsFromExecuteQueryResponse(e *proto.ExecuteQueryResponse) (interface{}, error) {
+	if er := e.GetE(); er != nil {
+		r, err := NewResultFromExecuteResult(er)
+		if err != nil {
+			return nil, err
+		}
+		return &ResultWithRows{
+			Result: *r,
+		}, nil
+	} else if qr := e.GetQ(); qr != nil {
+		return NewAssociativeRowsFromQueryRows(qr)
+	} else if err := e.GetError(); err != "" {
+		return map[string]string{
+			"error": err,
+		}, nil
+	}
+	return nil, errors.New("no ExecuteResult, QueryRows, or Error")
+}
+
 // NewResultFromExecuteResult returns an API Result object from an ExecuteResult.
-func NewResultFromExecuteResult(e *command.ExecuteResult) (*Result, error) {
+func NewResultFromExecuteResult(e *proto.ExecuteResult) (*Result, error) {
 	return &Result{
 		LastInsertID: e.LastInsertId,
 		RowsAffected: e.RowsAffected,
@@ -51,7 +93,7 @@ func NewResultFromExecuteResult(e *command.ExecuteResult) (*Result, error) {
 }
 
 // NewRowsFromQueryRows returns an API Rows object from a QueryRows
-func NewRowsFromQueryRows(q *command.QueryRows) (*Rows, error) {
+func NewRowsFromQueryRows(q *proto.QueryRows) (*Rows, error) {
 	if len(q.Columns) != len(q.Types) {
 		return nil, ErrTypesColumnsLengthViolation
 	}
@@ -70,7 +112,7 @@ func NewRowsFromQueryRows(q *command.QueryRows) (*Rows, error) {
 }
 
 // NewAssociativeRowsFromQueryRows returns an associative API object from a QueryRows
-func NewAssociativeRowsFromQueryRows(q *command.QueryRows) (*AssociativeRows, error) {
+func NewAssociativeRowsFromQueryRows(q *proto.QueryRows) (*AssociativeRows, error) {
 	if len(q.Columns) != len(q.Types) {
 		return nil, ErrTypesColumnsLengthViolation
 	}
@@ -103,7 +145,7 @@ func NewAssociativeRowsFromQueryRows(q *command.QueryRows) (*AssociativeRows, er
 }
 
 // NewValuesFromQueryValues sets Values from a QueryValue object.
-func NewValuesFromQueryValues(dest [][]interface{}, v []*command.Values) error {
+func NewValuesFromQueryValues(dest [][]interface{}, v []*proto.Values) error {
 	for n := range v {
 		vals := v[n]
 		if vals == nil {
@@ -120,15 +162,15 @@ func NewValuesFromQueryValues(dest [][]interface{}, v []*command.Values) error {
 		rowValues := make([]interface{}, len(params))
 		for p := range params {
 			switch w := params[p].GetValue().(type) {
-			case *command.Parameter_I:
+			case *proto.Parameter_I:
 				rowValues[p] = w.I
-			case *command.Parameter_D:
+			case *proto.Parameter_D:
 				rowValues[p] = w.D
-			case *command.Parameter_B:
+			case *proto.Parameter_B:
 				rowValues[p] = w.B
-			case *command.Parameter_Y:
+			case *proto.Parameter_Y:
 				rowValues[p] = w.Y
-			case *command.Parameter_S:
+			case *proto.Parameter_S:
 				rowValues[p] = w.S
 			case nil:
 				rowValues[p] = nil
@@ -142,7 +184,8 @@ func NewValuesFromQueryValues(dest [][]interface{}, v []*command.Values) error {
 	return nil
 }
 
-// Encoder is used to JSON marshal ExecuteResults and QueryRows
+// Encoder is used to JSON marshal ExecuteResults, QueryRows
+// and ExecuteQueryRequests.
 type Encoder struct {
 	Associative bool
 }
@@ -180,13 +223,13 @@ type marshalFunc func(i interface{}) ([]byte, error)
 
 func jsonMarshal(i interface{}, f marshalFunc, assoc bool) ([]byte, error) {
 	switch v := i.(type) {
-	case *command.ExecuteResult:
+	case *proto.ExecuteResult:
 		r, err := NewResultFromExecuteResult(v)
 		if err != nil {
 			return nil, err
 		}
 		return f(r)
-	case []*command.ExecuteResult:
+	case []*proto.ExecuteResult:
 		var err error
 		results := make([]*Result, len(v))
 		for j := range v {
@@ -196,7 +239,7 @@ func jsonMarshal(i interface{}, f marshalFunc, assoc bool) ([]byte, error) {
 			}
 		}
 		return f(results)
-	case *command.QueryRows:
+	case *proto.QueryRows:
 		if assoc {
 			r, err := NewAssociativeRowsFromQueryRows(v)
 			if err != nil {
@@ -210,7 +253,13 @@ func jsonMarshal(i interface{}, f marshalFunc, assoc bool) ([]byte, error) {
 			}
 			return f(r)
 		}
-	case []*command.QueryRows:
+	case *proto.ExecuteQueryResponse:
+		r, err := NewResultRowsFromExecuteQueryResponse(v)
+		if err != nil {
+			return nil, err
+		}
+		return f(r)
+	case []*proto.QueryRows:
 		var err error
 
 		if assoc {
@@ -232,7 +281,29 @@ func jsonMarshal(i interface{}, f marshalFunc, assoc bool) ([]byte, error) {
 			}
 			return f(rows)
 		}
-	case []*command.Values:
+	case []*proto.ExecuteQueryResponse:
+		if assoc {
+			res := make([]interface{}, len(v))
+			for j := range v {
+				r, err := NewAssociativeResultRowsFromExecuteQueryResponse(v[j])
+				if err != nil {
+					return nil, err
+				}
+				res[j] = r
+			}
+			return f(res)
+		} else {
+			res := make([]interface{}, len(v))
+			for j := range v {
+				r, err := NewResultRowsFromExecuteQueryResponse(v[j])
+				if err != nil {
+					return nil, err
+				}
+				res[j] = r
+			}
+			return f(res)
+		}
+	case []*proto.Values:
 		values := make([][]interface{}, len(v))
 		if err := NewValuesFromQueryValues(values, v); err != nil {
 			return nil, err
