@@ -2,9 +2,6 @@ package store
 
 // DBConfig represents the configuration of the underlying SQLite database.
 type DBConfig struct {
-	// Whether the database is in-memory only.
-	Memory bool `json:"memory"`
-
 	// SQLite on-disk path
 	OnDiskPath string `json:"on_disk_path,omitempty"`
 
@@ -13,6 +10,6 @@ type DBConfig struct {
 }
 
 // NewDBConfig returns a new DB config instance.
-func NewDBConfig(memory bool) *DBConfig {
-	return &DBConfig{Memory: memory}
+func NewDBConfig() *DBConfig {
+	return &DBConfig{}
 }
