@@ -3,30 +3,30 @@ package system
 import (
 	"fmt"
 	"net"
+	"os"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/rqlite/rqlite/cluster"
-	"github.com/rqlite/rqlite/db"
-	"github.com/rqlite/rqlite/http"
-	"github.com/rqlite/rqlite/queue"
-	"github.com/rqlite/rqlite/rtls"
-	"github.com/rqlite/rqlite/store"
-	"github.com/rqlite/rqlite/tcp"
+	"github.com/rqlite/rqlite/v8/cluster"
+	"github.com/rqlite/rqlite/v8/db"
+	"github.com/rqlite/rqlite/v8/http"
+	"github.com/rqlite/rqlite/v8/queue"
+	"github.com/rqlite/rqlite/v8/store"
+	"github.com/rqlite/rqlite/v8/tcp"
 )
 
 // Test_JoinLeaderNode tests a join operation between a leader and a new node.
 func Test_JoinLeaderNode(t *testing.T) {
-	leader := mustNewLeaderNode()
+	leader := mustNewLeaderNode("leader1")
 	defer leader.Deprovision()
 
-	node := mustNewNode(false)
-	defer node.Deprovision()
-	if err := node.Join(leader); err != nil {
+	node2 := mustNewNode("node1", false)
+	defer node2.Deprovision()
+	if err := node2.Join(leader); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
 	}
-	_, err := node.WaitForLeader()
+	_, err := node2.WaitForLeader()
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
@@ -34,10 +34,10 @@ func Test_JoinLeaderNode(t *testing.T) {
 
 // Test_MultiNodeCluster tests formation of a 3-node cluster, and its operation.
 func Test_MultiNodeCluster(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -63,7 +63,7 @@ func Test_MultiNodeCluster(t *testing.T) {
 		t.Fatalf("got incorrect number of followers: %d", len(followers))
 	}
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	if err := node3.Join(leader); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -170,10 +170,10 @@ func Test_MultiNodeCluster(t *testing.T) {
 // Test_MultiNodeClusterRANDOM tests operation of RANDOM() SQL rewriting. It checks that a rewritten
 // statement is sent to follower.
 func Test_MultiNodeClusterRANDOM(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -202,7 +202,7 @@ func Test_MultiNodeClusterRANDOM(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to query for count: %s", err.Error())
 	}
-	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[1]]}]}`; got != exp {
+	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[1]]}]}`; got != exp {
 		t.Fatalf("wrong query results, exp %s, got %s", exp, got)
 	}
 
@@ -213,39 +213,40 @@ func Test_MultiNodeClusterRANDOM(t *testing.T) {
 
 	// Check that row is *exactly* the same on each node. This could only happen if RANDOM was
 	// rewritten by the Leader before committing to the Raft log.
-	r1, err := node1.QueryNoneConsistency("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query node 1: %s", err.Error())
+	tFn := func() bool {
+		r1, err := node1.QueryNoneConsistency("SELECT * FROM foo")
+		if err != nil {
+			t.Fatalf("failed to query node 1: %s", err.Error())
+		}
+		r2, err := node2.QueryNoneConsistency("SELECT * FROM foo")
+		if err != nil {
+			t.Fatalf("failed to query node 2: %s", err.Error())
+		}
+		return r1 == r2
 	}
-	r2, err := node2.QueryNoneConsistency("SELECT * FROM foo")
-	if err != nil {
-		t.Fatalf("failed to query node 2: %s", err.Error())
-	}
-	if r1 != r2 {
-		t.Fatalf("node 1 and node 2 do not have the same data (%s %s)", r1, r2)
-	}
+	trueOrTimeout(tFn, 10*time.Second)
 }
 
 // Test_MultiNodeClusterBootstrap tests formation of a 3-node cluster via bootstraping,
 // and its operation.
 func Test_MultiNodeClusterBootstrap(t *testing.T) {
-	node1 := mustNewNode(false)
+	node1 := mustNewNode("node1", false)
 	node1.Store.BootstrapExpect = 3
 	defer node1.Deprovision()
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	node2.Store.BootstrapExpect = 3
 	defer node2.Deprovision()
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	node3.Store.BootstrapExpect = 3
 	defer node3.Deprovision()
 
 	provider := cluster.NewAddressProviderString(
-		[]string{node1.APIAddr, node2.APIAddr, node3.APIAddr})
-	node1Bs := cluster.NewBootstrapper(provider, nil)
-	node2Bs := cluster.NewBootstrapper(provider, nil)
-	node3Bs := cluster.NewBootstrapper(provider, nil)
+		[]string{node1.RaftAddr, node2.RaftAddr, node3.RaftAddr})
+	node1Bs := cluster.NewBootstrapper(provider, node1.Client)
+	node2Bs := cluster.NewBootstrapper(provider, node2.Client)
+	node3Bs := cluster.NewBootstrapper(provider, node3.Client)
 
 	// Have all nodes start a bootstrap basically in parallel,
 	// ensure only 1 leader actually gets elected.
@@ -256,7 +257,7 @@ func Test_MultiNodeClusterBootstrap(t *testing.T) {
 			addr, _ := node1.Store.LeaderAddr()
 			return addr != ""
 		}
-		node1Bs.Boot(node1.ID, node1.RaftAddr, done, 10*time.Second)
+		node1Bs.Boot(node1.ID, node1.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	go func() {
@@ -264,7 +265,7 @@ func Test_MultiNodeClusterBootstrap(t *testing.T) {
 			addr, _ := node2.Store.LeaderAddr()
 			return addr != ""
 		}
-		node2Bs.Boot(node2.ID, node2.RaftAddr, done, 10*time.Second)
+		node2Bs.Boot(node2.ID, node2.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	go func() {
@@ -272,7 +273,7 @@ func Test_MultiNodeClusterBootstrap(t *testing.T) {
 			addr, _ := node3.Store.LeaderAddr()
 			return addr != ""
 		}
-		node3Bs.Boot(node3.ID, node3.RaftAddr, done, 10*time.Second)
+		node3Bs.Boot(node3.ID, node3.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	wg.Wait()
@@ -394,25 +395,25 @@ func Test_MultiNodeClusterBootstrap(t *testing.T) {
 // Test_MultiNodeClusterBootstrapLaterJoin tests formation of a 3-node cluster and
 // then checking a 4th node can join later with the bootstap parameters.
 func Test_MultiNodeClusterBootstrapLaterJoin(t *testing.T) {
-	node1 := mustNewNode(false)
+	node1 := mustNewNode("node1", false)
 	node1.Store.BootstrapExpect = 3
 	defer node1.Deprovision()
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	node2.Store.BootstrapExpect = 3
 	defer node2.Deprovision()
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	node3.Store.BootstrapExpect = 3
 	defer node3.Deprovision()
 
 	provider := cluster.NewAddressProviderString(
-		[]string{node1.APIAddr, node2.APIAddr, node3.APIAddr})
-	node1Bs := cluster.NewBootstrapper(provider, nil)
+		[]string{node1.RaftAddr, node2.RaftAddr, node3.RaftAddr})
+	node1Bs := cluster.NewBootstrapper(provider, node1.Client)
 	node1Bs.Interval = time.Second
-	node2Bs := cluster.NewBootstrapper(provider, nil)
+	node2Bs := cluster.NewBootstrapper(provider, node2.Client)
 	node2Bs.Interval = time.Second
-	node3Bs := cluster.NewBootstrapper(provider, nil)
+	node3Bs := cluster.NewBootstrapper(provider, node3.Client)
 	node3Bs.Interval = time.Second
 
 	// Have all nodes start a bootstrap basically in parallel,
@@ -424,7 +425,7 @@ func Test_MultiNodeClusterBootstrapLaterJoin(t *testing.T) {
 			addr, _ := node1.Store.LeaderAddr()
 			return addr != ""
 		}
-		node1Bs.Boot(node1.ID, node1.RaftAddr, done, 10*time.Second)
+		node1Bs.Boot(node1.ID, node1.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	go func() {
@@ -432,7 +433,7 @@ func Test_MultiNodeClusterBootstrapLaterJoin(t *testing.T) {
 			addr, _ := node2.Store.LeaderAddr()
 			return addr != ""
 		}
-		node2Bs.Boot(node2.ID, node2.RaftAddr, done, 10*time.Second)
+		node2Bs.Boot(node2.ID, node2.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	go func() {
@@ -440,7 +441,7 @@ func Test_MultiNodeClusterBootstrapLaterJoin(t *testing.T) {
 			addr, _ := node3.Store.LeaderAddr()
 			return addr != ""
 		}
-		node3Bs.Boot(node3.ID, node3.RaftAddr, done, 10*time.Second)
+		node3Bs.Boot(node3.ID, node3.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	wg.Wait()
@@ -468,16 +469,16 @@ func Test_MultiNodeClusterBootstrapLaterJoin(t *testing.T) {
 
 	// Ensure a 4th node can join cluster with exactly same launch
 	// params. Under the cover it should just do a join.
-	node4 := mustNewNode(false)
+	node4 := mustNewNode("node4", false)
 	node4.Store.BootstrapExpect = 3
-	defer node3.Deprovision()
-	node4Bs := cluster.NewBootstrapper(provider, nil)
+	defer node4.Deprovision()
+	node4Bs := cluster.NewBootstrapper(provider, node4.Client)
 	node4Bs.Interval = time.Second
 	done := func() bool {
 		addr, _ := node4.Store.LeaderAddr()
 		return addr != ""
 	}
-	if err := node4Bs.Boot(node4.ID, node4.RaftAddr, done, 10*time.Second); err != nil {
+	if err := node4Bs.Boot(node4.ID, node4.RaftAddr, cluster.Voter, done, 10*time.Second); err != nil {
 		t.Fatalf("node 4 failed to boot")
 	}
 	node4Leader, err := node4.WaitForLeader()
@@ -489,33 +490,31 @@ func Test_MultiNodeClusterBootstrapLaterJoin(t *testing.T) {
 	}
 }
 
-// Test_MultiNodeClusterBootstrapLaterJoinHTTPS tests formation of a 3-node cluster which
+// Test_MultiNodeClusterBootstrapLaterJoinTLS tests formation of a 3-node cluster which
 // uses HTTP and TLS,then checking a 4th node can join later with the bootstap parameters.
-func Test_MultiNodeClusterBootstrapLaterJoinHTTPS(t *testing.T) {
-	node1 := mustNewNodeEncrypted(false, true, true)
+func Test_MultiNodeClusterBootstrapLaterJoinTLS(t *testing.T) {
+	node1 := mustNewNodeEncrypted("node1", false, true, true)
 	node1.Store.BootstrapExpect = 3
+	node1.EnableTLSClient()
 	defer node1.Deprovision()
 
-	node2 := mustNewNodeEncrypted(false, true, true)
+	node2 := mustNewNodeEncrypted("node2", false, true, true)
 	node2.Store.BootstrapExpect = 3
+	node2.EnableTLSClient()
 	defer node2.Deprovision()
 
-	node3 := mustNewNodeEncrypted(false, true, true)
+	node3 := mustNewNodeEncrypted("node3", false, true, true)
 	node3.Store.BootstrapExpect = 3
+	node3.EnableTLSClient()
 	defer node3.Deprovision()
 
-	tlsConfig, err := rtls.CreateClientConfig("", "", "", true, false)
-	if err != nil {
-		t.Fatalf("failed to create TLS config: %s", err)
-	}
-
 	provider := cluster.NewAddressProviderString(
-		[]string{node1.APIAddr, node2.APIAddr, node3.APIAddr})
-	node1Bs := cluster.NewBootstrapper(provider, tlsConfig)
+		[]string{node1.RaftAddr, node2.RaftAddr, node3.RaftAddr})
+	node1Bs := cluster.NewBootstrapper(provider, node1.Client)
 	node1Bs.Interval = time.Second
-	node2Bs := cluster.NewBootstrapper(provider, tlsConfig)
+	node2Bs := cluster.NewBootstrapper(provider, node2.Client)
 	node2Bs.Interval = time.Second
-	node3Bs := cluster.NewBootstrapper(provider, tlsConfig)
+	node3Bs := cluster.NewBootstrapper(provider, node3.Client)
 	node3Bs.Interval = time.Second
 
 	// Have all nodes start a bootstrap basically in parallel,
@@ -527,7 +526,7 @@ func Test_MultiNodeClusterBootstrapLaterJoinHTTPS(t *testing.T) {
 			addr, _ := node1.Store.LeaderAddr()
 			return addr != ""
 		}
-		node1Bs.Boot(node1.ID, node1.RaftAddr, done, 10*time.Second)
+		node1Bs.Boot(node1.ID, node1.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	go func() {
@@ -535,7 +534,7 @@ func Test_MultiNodeClusterBootstrapLaterJoinHTTPS(t *testing.T) {
 			addr, _ := node2.Store.LeaderAddr()
 			return addr != ""
 		}
-		node2Bs.Boot(node2.ID, node2.RaftAddr, done, 10*time.Second)
+		node2Bs.Boot(node2.ID, node2.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	go func() {
@@ -543,7 +542,7 @@ func Test_MultiNodeClusterBootstrapLaterJoinHTTPS(t *testing.T) {
 			addr, _ := node3.Store.LeaderAddr()
 			return addr != ""
 		}
-		node3Bs.Boot(node3.ID, node3.RaftAddr, done, 10*time.Second)
+		node3Bs.Boot(node3.ID, node3.RaftAddr, cluster.Voter, done, 10*time.Second)
 		wg.Done()
 	}()
 	wg.Wait()
@@ -570,17 +569,18 @@ func Test_MultiNodeClusterBootstrapLaterJoinHTTPS(t *testing.T) {
 	}
 
 	// Ensure a 4th node can join cluster with exactly same launch
-	// params. Under the cover it should just do a join.
-	node4 := mustNewNodeEncrypted(false, true, true)
+	// params. Under the covers it should just do a join.
+	node4 := mustNewNodeEncrypted("node4", false, true, true)
 	node4.Store.BootstrapExpect = 3
+	node4.EnableTLSClient()
 	defer node3.Deprovision()
-	node4Bs := cluster.NewBootstrapper(provider, tlsConfig)
+	node4Bs := cluster.NewBootstrapper(provider, node3.Client)
 	node4Bs.Interval = time.Second
 	done := func() bool {
 		addr, _ := node4.Store.LeaderAddr()
 		return addr != ""
 	}
-	if err := node4Bs.Boot(node4.ID, node4.RaftAddr, done, 10*time.Second); err != nil {
+	if err := node4Bs.Boot(node4.ID, node4.RaftAddr, cluster.Voter, done, 10*time.Second); err != nil {
 		t.Fatalf("node 4 failed to boot")
 	}
 	node4Leader, err := node4.WaitForLeader()
@@ -630,8 +630,11 @@ func Test_MultiNodeClusterRaftAdv(t *testing.T) {
 	}
 	go mux2.Serve()
 
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+
 	// Start two nodes, and ensure a cluster can be formed.
-	node1 := mustNodeEncrypted(mustTempDir(), true, false, mux1, "1")
+	node1 := mustNodeEncrypted("node1", mustTempDir("node1"), true, false, mux1, raftDialer, clstrDialer)
 	defer node1.Deprovision()
 	leader, err := node1.WaitForLeader()
 	if err != nil {
@@ -641,7 +644,7 @@ func Test_MultiNodeClusterRaftAdv(t *testing.T) {
 		t.Fatalf("node return wrong leader from leader, exp: %s, got %s", exp, got)
 	}
 
-	node2 := mustNodeEncrypted(mustTempDir(), false, false, mux2, "2")
+	node2 := mustNodeEncrypted("node2", mustTempDir("node2"), false, false, mux2, raftDialer, clstrDialer)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node2 failed to join leader: %s", err.Error())
@@ -657,10 +660,10 @@ func Test_MultiNodeClusterRaftAdv(t *testing.T) {
 
 // Test_MultiNodeClusterNodes checks nodes/ endpoint under various situations.
 func Test_MultiNodeClusterNodes(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -677,7 +680,7 @@ func Test_MultiNodeClusterNodes(t *testing.T) {
 		t.Fatalf("failed to find cluster leader: %s", err.Error())
 	}
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	if err := node3.Join(leader); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -702,8 +705,8 @@ func Test_MultiNodeClusterNodes(t *testing.T) {
 	if len(nodes) != len(c) {
 		t.Fatalf("nodes/ output returned wrong number of nodes, got %d, exp %d", len(nodes), len(c))
 	}
-	ns, ok := nodes[leader.ID]
-	if !ok {
+	ns := nodes.GetNode(leader.ID)
+	if ns == nil {
 		t.Fatalf("failed to find leader with ID %s in node status", leader.ID)
 	}
 	if !ns.Leader {
@@ -729,7 +732,10 @@ func Test_MultiNodeClusterNodes(t *testing.T) {
 		t.Fatalf("got incorrect number of followers: %d", len(followers))
 	}
 	f := followers[0]
-	ns = nodes[f.ID]
+	ns = nodes.GetNode(f.ID)
+	if ns == nil {
+		t.Fatalf("failed to find follower with ID %s in node status", f.ID)
+	}
 	if ns.Addr != f.RaftAddr {
 		t.Fatalf("node has wrong Raft address for follower")
 	}
@@ -747,7 +753,7 @@ func Test_MultiNodeClusterNodes(t *testing.T) {
 // Test_MultiNodeClusterQueuedWrites tests writing to a cluster using
 // normal and queued writes.
 func Test_MultiNodeClusterQueuedWrites(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
 	if _, err := node1.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`); err != nil {
@@ -755,7 +761,7 @@ func Test_MultiNodeClusterQueuedWrites(t *testing.T) {
 	}
 
 	// Join a second and third nodes
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -764,7 +770,7 @@ func Test_MultiNodeClusterQueuedWrites(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	if err := node3.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -777,12 +783,13 @@ func Test_MultiNodeClusterQueuedWrites(t *testing.T) {
 	// Write data to the cluster, via various methods and nodes.
 	writesPerLoop := 500
 	var wg sync.WaitGroup
-	wg.Add(5)
+	numLoops := 5
+	wg.Add(numLoops)
 	go func() {
 		defer wg.Done()
 		for i := 0; i < writesPerLoop; i++ {
 			if _, err := node1.Execute(`INSERT INTO foo(name) VALUES("fiona")`); err != nil {
-				t.Logf("failed to insert records: %s", err.Error())
+				t.Errorf("failed to insert records: %s", err.Error())
 			}
 		}
 	}()
@@ -790,7 +797,7 @@ func Test_MultiNodeClusterQueuedWrites(t *testing.T) {
 		defer wg.Done()
 		for i := 0; i < writesPerLoop; i++ {
 			if _, err := node2.Execute(`INSERT INTO foo(name) VALUES("fiona")`); err != nil {
-				t.Logf("failed to insert records: %s", err.Error())
+				t.Errorf("failed to insert records: %s", err.Error())
 			}
 		}
 	}()
@@ -798,38 +805,38 @@ func Test_MultiNodeClusterQueuedWrites(t *testing.T) {
 		defer wg.Done()
 		for i := 0; i < writesPerLoop-1; i++ {
 			if _, err := node2.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, false); err != nil {
-				t.Logf("failed to insert records: %s", err.Error())
+				t.Errorf("failed to insert records: %s\n", err.Error())
 			}
 		}
 		if _, err := node2.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, true); err != nil {
-			t.Logf("failed to insert records: %s", err.Error())
+			t.Errorf("failed to insert records: %s\n", err.Error())
 		}
 	}()
 	go func() {
 		defer wg.Done()
 		for i := 0; i < writesPerLoop-1; i++ {
 			if _, err := node3.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, false); err != nil {
-				t.Logf("failed to insert records: %s", err.Error())
+				t.Errorf("failed to insert records: %s\n", err.Error())
 			}
 		}
 		if _, err := node3.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, true); err != nil {
-			t.Logf("failed to insert records: %s", err.Error())
+			t.Errorf("failed to insert records: %s\n", err.Error())
 		}
 	}()
 	go func() {
 		defer wg.Done()
 		for i := 0; i < writesPerLoop-1; i++ {
 			if _, err := node3.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, false); err != nil {
-				t.Logf("failed to insert records: %s", err.Error())
+				t.Errorf("failed to insert records: %s\n", err.Error())
 			}
 		}
 		if _, err := node3.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, true); err != nil {
-			t.Logf("failed to insert records: %s", err.Error())
+			t.Errorf("failed to insert records: %s", err.Error())
 		}
 	}()
 	wg.Wait()
 
-	exp := fmt.Sprintf(`{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[%d]]}]}`, 5*writesPerLoop)
+	exp := fmt.Sprintf(`{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[%d]]}]}`, numLoops*writesPerLoop)
 	got, err := node1.Query(`SELECT COUNT(*) FROM foo`)
 	if err != nil {
 		t.Fatalf("failed to query follower node: %s", err.Error())
@@ -847,7 +854,7 @@ func Test_MultiNodeClusterLargeQueuedWrites(t *testing.T) {
 	http.ResetStats()
 	queue.ResetStats()
 
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
 	if _, err := node1.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`); err != nil {
@@ -855,7 +862,7 @@ func Test_MultiNodeClusterLargeQueuedWrites(t *testing.T) {
 	}
 
 	// Join a second and third nodes
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -864,7 +871,7 @@ func Test_MultiNodeClusterLargeQueuedWrites(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	if err := node3.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -876,26 +883,30 @@ func Test_MultiNodeClusterLargeQueuedWrites(t *testing.T) {
 
 	// Write data to the cluster, via various nodes.
 	nodesUnderTest := []*Node{node3, node1, node2, node1, node2, node3, node1, node3, node2}
-	writesPerNode := 10000
+	writesPerNode := 1000
+	writeBandBase := writesPerNode * 10
 
 	var wg sync.WaitGroup
 	wg.Add(len(nodesUnderTest))
-	for _, n := range nodesUnderTest {
-		go func(nt *Node) {
+	for i, n := range nodesUnderTest {
+		go func(ii int, nt *Node) {
 			defer wg.Done()
-			for i := 0; i < writesPerNode-1; i++ {
-				if _, err := nt.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, false); err != nil {
+			var j int
+			for j = 0; j < writesPerNode-1; j++ {
+				stmt := fmt.Sprintf(`INSERT OR IGNORE INTO foo(id, name) VALUES(%d, "fiona")`, j+(ii*writeBandBase))
+				if _, err := nt.ExecuteQueued(stmt, false); err != nil {
 					t.Logf("failed to insert records: %s", err.Error())
 				}
 			}
-			if _, err := nt.ExecuteQueued(`INSERT INTO foo(name) VALUES("fiona")`, true); err != nil {
+			stmt := fmt.Sprintf(`INSERT OR IGNORE INTO foo(id, name) VALUES(%d, "fiona")`, j+(ii*writeBandBase))
+			if _, err := nt.ExecuteQueued(stmt, true); err != nil {
 				t.Logf("failed to insert records: %s", err.Error())
 			}
-		}(n)
+		}(i, n)
 	}
 	wg.Wait()
 
-	exp := fmt.Sprintf(`{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[%d]]}]}`, len(nodesUnderTest)*writesPerNode)
+	exp := fmt.Sprintf(`{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[%d]]}]}`, len(nodesUnderTest)*writesPerNode)
 	got, err := node1.Query(`SELECT COUNT(*) FROM foo`)
 	if err != nil {
 		t.Fatalf("failed to query follower node: %s", err.Error())
@@ -907,10 +918,10 @@ func Test_MultiNodeClusterLargeQueuedWrites(t *testing.T) {
 
 // Test_MultiNodeClusterNodesNonVoter checks nodes/ endpoint with a non-voting node.
 func Test_MultiNodeClusterNodesNonVoter(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -927,7 +938,7 @@ func Test_MultiNodeClusterNodesNonVoter(t *testing.T) {
 		t.Fatalf("failed to find cluster leader: %s", err.Error())
 	}
 
-	nonVoter := mustNewNode(false)
+	nonVoter := mustNewNode("nonvoter", false)
 	defer nonVoter.Deprovision()
 	if err := nonVoter.JoinAsNonVoter(leader); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -978,13 +989,15 @@ func Test_MultiNodeClusterNodesNonVoter(t *testing.T) {
 // Test_MultiNodeClusterNodeEncrypted tests formation of a 3-node cluster, and its operation.
 // This test enables inter-node encryption, but keeps the unencrypted HTTP API.
 func Test_MultiNodeClusterNodeEncrypted(t *testing.T) {
-	node1 := mustNewNodeEncrypted(true, false, true)
+	node1 := mustNewNodeEncrypted("node1", true, false, true)
+	node1.EnableTLSClient()
 	defer node1.Deprovision()
 	if _, err := node1.WaitForLeader(); err != nil {
 		t.Fatalf("node never became leader")
 	}
 
-	node2 := mustNewNodeEncrypted(false, false, true)
+	node2 := mustNewNodeEncrypted("node2", false, false, true)
+	node2.EnableTLSClient()
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1010,7 +1023,8 @@ func Test_MultiNodeClusterNodeEncrypted(t *testing.T) {
 		t.Fatalf("got incorrect number of followers: %d", len(followers))
 	}
 
-	node3 := mustNewNodeEncrypted(false, false, true)
+	node3 := mustNewNodeEncrypted("node3", false, false, true)
+	node3.EnableTLSClient()
 	defer node3.Deprovision()
 	if err := node3.Join(leader); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1125,23 +1139,48 @@ func Test_MultiNodeClusterNodeEncrypted(t *testing.T) {
 
 // Test_MultiNodeClusterSnapshot tests formation of a 3-node cluster, which involves sharing snapshots.
 func Test_MultiNodeClusterSnapshot(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
 	if _, err := node1.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`); err != nil {
 		t.Fatalf("failed to create table: %s", err.Error())
 	}
 
-	// Force snapshots and log truncation to occur.
-	for i := 0; i < 3*int(node1.Store.SnapshotThreshold); i++ {
+	// Force snapshots and log truncation to occur. Make the number of snapshots high enough
+	// that some logs are deleted from the log and are not present in trailing logs either.
+	// This way we can be sure that the follower will need to get an actual snapshot from the leader.
+	for i := 0; i < 5*int(node1.Store.SnapshotThreshold); i++ {
 		_, err := node1.Execute(`INSERT INTO foo(name) VALUES("sinead")`)
 		if err != nil {
 			t.Fatalf(`failed to write records for Snapshot test: %s`, err.Error())
 		}
 	}
 
-	// Join a second and third nodes, which will get database state via snapshots.
-	node2 := mustNewNode(false)
+	expResults := `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[500]]}]}`
+	testerFn := func(n *Node) {
+		c := 0
+		for {
+			r, err := n.QueryNoneConsistency(`SELECT COUNT(*) FROM foo`)
+			if err != nil {
+				t.Fatalf("failed to query follower node: %s", err.Error())
+			}
+
+			if r != expResults {
+				if c < 10 {
+					// Wait, and try again.
+					sleepForSecond()
+					c++
+					continue
+				}
+				t.Fatalf("timed out waiting for snapshot state")
+			}
+			// The node passed!
+			break
+		}
+	}
+
+	// Join a second node, check it gets the data via a snapshot.
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1150,8 +1189,10 @@ func Test_MultiNodeClusterSnapshot(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
+	testerFn(node2)
 
-	node3 := mustNewNode(false)
+	// Create and add a third node to the cluster.
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	if err := node3.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1160,8 +1201,6 @@ func Test_MultiNodeClusterSnapshot(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
-
-	// Create a new cluster.
 	c := Cluster{node1, node2, node3}
 
 	// Wait for followers to pick up state.
@@ -1170,28 +1209,9 @@ func Test_MultiNodeClusterSnapshot(t *testing.T) {
 		t.Fatalf("failed to determine followers: %s", err.Error())
 	}
 
-	var n int
-	var r string
+	// Check that all followers have the correct state.
 	for _, f := range followers {
-		n = 0
-		for {
-			r, err = f.QueryNoneConsistency(`SELECT COUNT(*) FROM foo`)
-			if err != nil {
-				t.Fatalf("failed to query follower node: %s", err.Error())
-			}
-
-			if r != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[300]]}]}` {
-				if n < 20 {
-					// Wait, and try again.
-					time.Sleep(mustParseDuration("1s"))
-					n++
-					continue
-				}
-				t.Fatalf("timed out waiting for snapshot state")
-			}
-			// The node passed!
-			break
-		}
+		testerFn(f)
 	}
 
 	// Kill original node.
@@ -1204,35 +1224,70 @@ func Test_MultiNodeClusterSnapshot(t *testing.T) {
 	}
 
 	// Test that the new leader still has the full state.
-	n = 0
-	for {
-		var r string
-		r, err = leader.Query(`SELECT COUNT(*) FROM foo`)
-		if err != nil {
-			t.Fatalf("failed to query follower node: %s", err.Error())
-		}
+	testerFn(leader)
+}
 
-		if r != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[300]]}]}` {
-			if n < 10 {
-				// Wait, and try again.
-				time.Sleep(mustParseDuration("100ms"))
-				n++
-				continue
-			}
-			t.Fatalf("timed out waiting for snapshot state")
-		}
-		// Test passed!
-		break
+func Test_MultiNodeCluster_Backup(t *testing.T) {
+	node1 := mustNewLeaderNode("leader1")
+	defer node1.Deprovision()
+
+	node2 := mustNewNode("node2", false)
+	defer node2.Deprovision()
+	if err := node2.Join(node1); err != nil {
+		t.Fatalf("node failed to join leader: %s", err.Error())
+	}
+	_, err := node2.WaitForLeader()
+	if err != nil {
+		t.Fatalf("failed waiting for leader: %s", err.Error())
+	}
+
+	// Create a table and write a record
+	if _, err := node1.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`); err != nil {
+		t.Fatalf("failed to create table: %s", err.Error())
+	}
+	if _, err := node1.Execute(`INSERT INTO foo(name) VALUES("fiona")`); err != nil {
+		t.Fatalf("failed to insert record: %s", err.Error())
+	}
+
+	// Get a backup from the leader via the follower.
+	c := Cluster{node1, node2}
+	followers, err := c.Followers()
+	if err != nil {
+		t.Fatalf("failed to get followers: %s", err.Error())
+	}
+
+	backupFile := mustTempFile()
+	defer os.Remove(backupFile)
+	if err := followers[0].Backup(backupFile, false); err != nil {
+		t.Fatalf("failed to get backup from follower: %s", err.Error())
+	}
+
+	// Create brand new node, and restore backup into it.
+	newNode := mustNewLeaderNode("newNode")
+	defer newNode.Deprovision()
+	_, err = newNode.Boot(backupFile)
+	if err != nil {
+		t.Fatalf("failed to boot using backup: %s", err.Error())
+	}
+
+	// Check that the new node has the correct state.
+	exp := `{"results":[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]}`
+	got, err := newNode.Query(`SELECT * FROM foo`)
+	if err != nil {
+		t.Fatalf("failed to query new node: %s", err.Error())
+	}
+	if got != exp {
+		t.Fatalf("incorrect count, got %s, exp %s", got, exp)
 	}
 }
 
 // Test_MultiNodeClusterWithNonVoter tests formation of a 4-node cluster, one of which is
 // a non-voter
 func Test_MultiNodeClusterWithNonVoter(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1249,7 +1304,7 @@ func Test_MultiNodeClusterWithNonVoter(t *testing.T) {
 		t.Fatalf("failed to find cluster leader: %s", err.Error())
 	}
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	if err := node3.Join(leader); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1266,7 +1321,7 @@ func Test_MultiNodeClusterWithNonVoter(t *testing.T) {
 		t.Fatalf("failed to find cluster leader: %s", err.Error())
 	}
 
-	nonVoter := mustNewNode(false)
+	nonVoter := mustNewNode("nonvoter", false)
 	defer nonVoter.Deprovision()
 	if err := nonVoter.JoinAsNonVoter(leader); err != nil {
 		t.Fatalf("non-voting node failed to join leader: %s", err.Error())
@@ -1367,7 +1422,7 @@ func Test_MultiNodeClusterWithNonVoter(t *testing.T) {
 // Test_MultiNodeClusterRecoverSingle tests recovery of a single node from a 3-node cluster,
 // which no longer has quorum.
 func Test_MultiNodeClusterRecoverSingle(t *testing.T) {
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 
 	if _, err := node1.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`); err != nil {
@@ -1376,12 +1431,12 @@ func Test_MultiNodeClusterRecoverSingle(t *testing.T) {
 	if _, err := node1.Execute(`INSERT INTO foo(id, name) VALUES(1, "fiona")`); err != nil {
 		t.Fatalf("failed to create table: %s", err.Error())
 	}
-	if rows, _ := node1.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[1]]}]}` {
+	if rows, _ := node1.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[1]]}]}` {
 		t.Fatalf("got incorrect results from node: %s", rows)
 	}
 
 	// Join a second and third nodes, which will get database state via snapshots.
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1390,8 +1445,11 @@ func Test_MultiNodeClusterRecoverSingle(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
+	if rows, _ := node2.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[1]]}]}` {
+		t.Fatalf("got incorrect results from node: %s", rows)
+	}
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	if err := node3.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
@@ -1399,6 +1457,9 @@ func Test_MultiNodeClusterRecoverSingle(t *testing.T) {
 	_, err = node3.WaitForLeader()
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
+	}
+	if rows, _ := node3.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[1]]}]}` {
+		t.Fatalf("got incorrect results from node: %s", rows)
 	}
 
 	// Shutdown all nodes
@@ -1412,11 +1473,14 @@ func Test_MultiNodeClusterRecoverSingle(t *testing.T) {
 		t.Fatalf("failed to close node3: %s", err.Error())
 	}
 
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+
 	// Create a single node using the node's data directory. It should fail because
 	// quorum can't be met. This isn't quite right since the Raft address is also
 	// changing, but it generally proves it doesn't come up.
 	mux0, ln0 := mustNewOpenMux("127.0.0.1:10000")
-	failedSingle := mustNodeEncrypted(node1.Dir, true, false, mux0, node1.Store.ID())
+	failedSingle := mustNodeEncrypted(node1.Store.ID(), node1.Dir, true, false, mux0, raftDialer, clstrDialer)
 	_, err = failedSingle.WaitForLeader()
 	if err == nil {
 		t.Fatalf("no error waiting for leader")
@@ -1429,12 +1493,12 @@ func Test_MultiNodeClusterRecoverSingle(t *testing.T) {
 	peers := fmt.Sprintf(`[{"id": "%s","address": "%s"}]`, node1.Store.ID(), "127.0.0.1:10001")
 	mustWriteFile(node1.PeersPath, peers)
 
-	okSingle := mustNodeEncrypted(node1.Dir, true, false, mux1, node1.Store.ID())
+	okSingle := mustNodeEncrypted(node1.Store.ID(), node1.Dir, true, false, mux1, raftDialer, clstrDialer)
 	_, err = okSingle.WaitForLeader()
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
-	if rows, _ := okSingle.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[1]]}]}` {
+	if rows, _ := okSingle.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[1]]}]}` {
 		t.Fatalf("got incorrect results from recovered node: %s", rows)
 	}
 	okSingle.Close(true)
@@ -1446,15 +1510,18 @@ func Test_MultiNodeClusterRecoverSingle(t *testing.T) {
 func Test_MultiNodeClusterRecoverFull(t *testing.T) {
 	var err error
 
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+
 	mux1, ln1 := mustNewOpenMux("127.0.0.1:10001")
-	node1 := mustNodeEncrypted(mustTempDir(), true, false, mux1, "1")
+	node1 := mustNodeEncrypted("node1", mustTempDir("node1"), true, false, mux1, raftDialer, clstrDialer)
 	_, err = node1.WaitForLeader()
 	if err != nil {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
 
 	mux2, ln2 := mustNewOpenMux("127.0.0.1:10002")
-	node2 := mustNodeEncrypted(mustTempDir(), false, false, mux2, "2")
+	node2 := mustNodeEncrypted("node2", mustTempDir("node2"), false, false, mux2, raftDialer, clstrDialer)
 	if err := node2.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
 	}
@@ -1464,7 +1531,7 @@ func Test_MultiNodeClusterRecoverFull(t *testing.T) {
 	}
 
 	mux3, ln3 := mustNewOpenMux("127.0.0.1:10003")
-	node3 := mustNodeEncrypted(mustTempDir(), false, false, mux3, "3")
+	node3 := mustNodeEncrypted("node3", mustTempDir("node3"), false, false, mux3, raftDialer, clstrDialer)
 	if err := node3.Join(node1); err != nil {
 		t.Fatalf("node failed to join leader: %s", err.Error())
 	}
@@ -1479,8 +1546,12 @@ func Test_MultiNodeClusterRecoverFull(t *testing.T) {
 	if _, err := node1.Execute(`INSERT INTO foo(id, name) VALUES(1, "fiona")`); err != nil {
 		t.Fatalf("failed to create table: %s", err.Error())
 	}
-	if rows, _ := node1.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[1]]}]}` {
-		t.Fatalf("got incorrect results from node: %s", rows)
+	rows, err := node1.Query(`SELECT COUNT(*) FROM foo`)
+	if err != nil {
+		t.Fatalf("failed to query node: %s", err.Error())
+	}
+	if got, exp := rows, `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[1]]}]}`; got != exp {
+		t.Fatalf("got incorrect results from node exp: %s got: %s", exp, got)
 	}
 
 	// Shutdown all nodes
@@ -1508,27 +1579,33 @@ func Test_MultiNodeClusterRecoverFull(t *testing.T) {
 	mustWriteFile(node3.PeersPath, peers)
 
 	mux4, ln4 := mustNewOpenMux("127.0.0.1:11001")
-	node4 := mustNodeEncrypted(node1.Dir, false, false, mux4, "1")
+	node4 := mustNodeEncrypted("1", node1.Dir, false, false, mux4, raftDialer, clstrDialer)
 	defer node4.Deprovision()
 	defer ln4.Close()
 
 	mux5, ln5 := mustNewOpenMux("127.0.0.1:11002")
-	node5 := mustNodeEncrypted(node2.Dir, false, false, mux5, "2")
+	node5 := mustNodeEncrypted("2", node2.Dir, false, false, mux5, raftDialer, clstrDialer)
 	defer node5.Deprovision()
 	defer ln5.Close()
 
 	mux6, ln6 := mustNewOpenMux("127.0.0.1:11003")
-	node6 := mustNodeEncrypted(node3.Dir, false, false, mux6, "3")
+	node6 := mustNodeEncrypted("3", node3.Dir, false, false, mux6, raftDialer, clstrDialer)
 	defer node6.Deprovision()
 	defer ln6.Close()
 
-	_, err = node6.WaitForLeader()
-	if err != nil {
-		t.Fatalf("failed waiting for leader on recovered cluster: %s", err.Error())
+	for _, node := range []*Node{node4, node5, node6} {
+		_, err = node.WaitForLeader()
+		if err != nil {
+			t.Fatalf("failed waiting for leader on node %s (recovered cluster): %s", node.ID, err.Error())
+		}
 	}
 
-	if rows, _ := node4.Query(`SELECT COUNT(*) FROM foo`); rows != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[1]]}]}` {
-		t.Fatalf("got incorrect results from recovered node: %s", rows)
+	rows, err = node4.Query(`SELECT COUNT(*) FROM foo`)
+	if err != nil {
+		t.Fatalf("failed to query recovered node: %s", err.Error())
+	}
+	if got, exp := rows, `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[1]]}]}`; got != exp {
+		t.Fatalf("got incorrect results from recovered node exp: %s got: %s", exp, got)
 	}
 }
 
@@ -1539,7 +1616,7 @@ func Test_MultiNodeClusterReapNodes(t *testing.T) {
 		n.Store.ReapReadOnlyTimeout = time.Second
 	}
 
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 	cfgStoreFn(node1)
 	_, err := node1.WaitForLeader()
@@ -1547,7 +1624,7 @@ func Test_MultiNodeClusterReapNodes(t *testing.T) {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	cfgStoreFn(node2)
 	if err := node2.Join(node1); err != nil {
@@ -1565,7 +1642,7 @@ func Test_MultiNodeClusterReapNodes(t *testing.T) {
 		t.Fatalf("failed to find cluster leader: %s", err.Error())
 	}
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	cfgStoreFn(node3)
 	if err := node3.Join(leader); err != nil {
@@ -1583,7 +1660,7 @@ func Test_MultiNodeClusterReapNodes(t *testing.T) {
 		t.Fatalf("failed to find cluster leader: %s", err.Error())
 	}
 
-	nonVoter := mustNewNode(false)
+	nonVoter := mustNewNode("nonvoter", false)
 	defer nonVoter.Deprovision()
 	cfgStoreFn(nonVoter)
 	if err := nonVoter.JoinAsNonVoter(leader); err != nil {
@@ -1641,7 +1718,7 @@ func Test_MultiNodeClusterNoReap(t *testing.T) {
 		n.Store.ReapReadOnlyTimeout = 120 * time.Second
 	}
 
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 	cfgStoreFn(node1)
 	_, err := node1.WaitForLeader()
@@ -1649,7 +1726,7 @@ func Test_MultiNodeClusterNoReap(t *testing.T) {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
 
-	nonVoter := mustNewNode(false)
+	nonVoter := mustNewNode("nonvoter", false)
 	defer nonVoter.Deprovision()
 	cfgStoreFn(nonVoter)
 	if err := nonVoter.JoinAsNonVoter(node1); err != nil {
@@ -1686,7 +1763,7 @@ func Test_MultiNodeClusterNoReapZero(t *testing.T) {
 		n.Store.ReapTimeout = 0
 	}
 
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 	cfgStoreFn(node1)
 	_, err := node1.WaitForLeader()
@@ -1694,7 +1771,7 @@ func Test_MultiNodeClusterNoReapZero(t *testing.T) {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
 
-	node2 := mustNewNode(false)
+	node2 := mustNewNode("node2", false)
 	defer node2.Deprovision()
 	cfgStoreFn(node2)
 	if err := node2.Join(node1); err != nil {
@@ -1712,7 +1789,7 @@ func Test_MultiNodeClusterNoReapZero(t *testing.T) {
 		t.Fatalf("failed to find cluster leader: %s", err.Error())
 	}
 
-	node3 := mustNewNode(false)
+	node3 := mustNewNode("node3", false)
 	defer node3.Deprovision()
 	cfgStoreFn(node3)
 	if err := node3.Join(leader); err != nil {
@@ -1757,7 +1834,7 @@ func Test_MultiNodeClusterNoReapReadOnlyZero(t *testing.T) {
 		n.Store.ReapReadOnlyTimeout = 0
 	}
 
-	node1 := mustNewLeaderNode()
+	node1 := mustNewLeaderNode("leader1")
 	defer node1.Deprovision()
 	cfgStoreFn(node1)
 	_, err := node1.WaitForLeader()
@@ -1765,7 +1842,7 @@ func Test_MultiNodeClusterNoReapReadOnlyZero(t *testing.T) {
 		t.Fatalf("failed waiting for leader: %s", err.Error())
 	}
 
-	nonVoter := mustNewNode(false)
+	nonVoter := mustNewNode("2", false)
 	defer nonVoter.Deprovision()
 	cfgStoreFn(nonVoter)
 	if err := nonVoter.JoinAsNonVoter(node1); err != nil {
@@ -1794,4 +1871,52 @@ func Test_MultiNodeClusterNoReapReadOnlyZero(t *testing.T) {
 	if trueOrTimeout(tFn, 10*time.Second) {
 		t.Fatalf("didn't time out waiting for node to be removed")
 	}
+}
+
+func Test_MultiNodeCluster_Boot(t *testing.T) {
+	node1 := mustNewLeaderNode("leader1")
+	defer node1.Deprovision()
+
+	_, err := node1.Boot("testdata/auto-restore.sqlite")
+	if err != nil {
+		t.Fatalf("failed to boot: %s", err.Error())
+	}
+
+	// Join a second node, check it gets the data via a snapshot.
+	node2 := mustNewNode("node2", false)
+	defer node2.Deprovision()
+	if err := node2.Join(node1); err != nil {
+		t.Fatalf("node failed to join leader: %s", err.Error())
+	}
+	_, err = node2.WaitForLeader()
+	if err != nil {
+		t.Fatalf("failed waiting for leader: %s", err.Error())
+	}
+
+	// Send a few Noops through to ensure SQLite database has been updated on each node.
+	for i := 0; i < 5; i++ {
+		node1.Noop("some_id")
+	}
+
+	// Check the database on each node
+	for _, n := range []*Node{node1, node2} {
+		rows, err := n.QueryNoneConsistency(`SELECT COUNT(*) FROM foo`)
+		if err != nil {
+			t.Fatalf("failed to query node: %s", err.Error())
+		}
+		if got, exp := rows, `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[3]]}]}`; got != exp {
+			t.Fatalf("got incorrect results from node exp: %s got: %s", exp, got)
+		}
+	}
+
+	// One last test -- the leader should refuse any more Boot requests because it
+	// is now part of a cluster.
+	_, err = node1.Boot("testdata/auto-restore.sqlite")
+	if err == nil {
+		t.Fatalf("expected error booting")
+	}
+}
+
+func sleepForSecond() {
+	time.Sleep(mustParseDuration("1s"))
 }
