@@ -4,7 +4,9 @@ Package system runs system-level testing of rqlite. This includes testing of sin
 package system
 
 import (
+	"compress/gzip"
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -12,29 +14,34 @@ import (
 	"testing"
 	"time"
 
-	"github.com/rqlite/rqlite/store"
+	"github.com/rqlite/rqlite/v8/cluster"
+	httpd "github.com/rqlite/rqlite/v8/http"
+	"github.com/rqlite/rqlite/v8/store"
+	"github.com/rqlite/rqlite/v8/tcp"
 )
 
 func Test_SingleNodeBasicEndpoint(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("node1")
 	defer node.Deprovision()
 
 	// Ensure accessing endpoints in basic manner works
 	_, err := node.Status()
 	if err != nil {
-		t.Fatalf(`failed to retrieve status for in-memory: %s`, err)
+		t.Fatalf(`failed to retrieve status: %s`, err)
 	}
 
-	dir := mustTempDir()
+	dir := mustTempDir("node1")
 	mux, ln := mustNewOpenMux("")
 	defer ln.Close()
-	node = mustNodeEncryptedOnDisk(dir, true, false, mux, "", false)
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+	node = mustNodeEncrypted("node1", dir, true, false, mux, raftDialer, clstrDialer)
 	if _, err := node.WaitForLeader(); err != nil {
 		t.Fatalf("node never became leader")
 	}
 	_, err = node.Status()
 	if err != nil {
-		t.Fatalf(`failed to retrieve status for on-disk: %s`, err)
+		t.Fatalf(`failed to retrieve status for: %s`, err)
 	}
 
 	ready, err := node.Ready()
@@ -44,10 +51,23 @@ func Test_SingleNodeBasicEndpoint(t *testing.T) {
 	if !ready {
 		t.Fatalf("node is not ready")
 	}
+
+	// Test that registering a ready channel affects readiness as expected.
+	ch := make(chan struct{})
+	node.Store.RegisterReadyChannel(ch)
+	ready, err = node.Ready()
+	if err != nil {
+		t.Fatalf(`failed to retrieve readiness: %s`, err)
+	}
+	if ready {
+		t.Fatalf("node is ready when registered ready channel not yet closed")
+	}
+	close(ch)
+	testPoll(t, node.Ready, 100*time.Millisecond, 5*time.Second)
 }
 
 func Test_SingleNodeNotReadyLive(t *testing.T) {
-	node := mustNewNode(false)
+	node := mustNewNode("node1", false)
 	defer node.Deprovision()
 	ready, err := node.Ready()
 	if err != nil {
@@ -78,7 +98,7 @@ func Test_SingleNodeNotReadyLive(t *testing.T) {
 }
 
 func Test_SingleNode(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("node1")
 	defer node.Deprovision()
 
 	tests := []struct {
@@ -140,8 +160,57 @@ func Test_SingleNode(t *testing.T) {
 	}
 }
 
+func Test_SingleNodeRequest(t *testing.T) {
+	node := mustNewLeaderNode("leader1")
+	defer node.Deprovision()
+
+	tests := []struct {
+		stmt     string
+		expected string
+	}{
+		{
+			stmt:     `CREATE TABLE foo (id integer not null primary key, name text)`,
+			expected: `{"results":[{}]}`,
+		},
+		{
+			stmt:     `INSERT INTO foo(name) VALUES("fiona")`,
+			expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
+		},
+		{
+			stmt:     `INSERT INTO bar(name) VALUES("fiona")`,
+			expected: `{"results":[{"error":"no such table: bar"}]}`,
+		},
+		{
+			stmt:     `INSERT blah blah`,
+			expected: `{"results":[{"error":"near \"blah\": syntax error"}]}`,
+		},
+		{
+			stmt:     `SELECT * FROM foo`,
+			expected: `{"results":[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]}`,
+		},
+		{
+			stmt:     `DROP TABLE bar`,
+			expected: `{"results":[{"error":"no such table: bar"}]}`,
+		},
+		{
+			stmt:     `DROP TABLE foo`,
+			expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
+		},
+	}
+
+	for i, tt := range tests {
+		r, err := node.Request(tt.stmt)
+		if err != nil {
+			t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
+		}
+		if r != tt.expected {
+			t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
+		}
+	}
+}
+
 func Test_SingleNodeMulti(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	tests := []struct {
@@ -203,7 +272,7 @@ func Test_SingleNodeMulti(t *testing.T) {
 
 func Test_SingleNodeConcurrentRequests(t *testing.T) {
 	var err error
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	node.Store.SetRequestCompression(1024, 1024) // Ensure no compression
 	defer node.Deprovision()
 
@@ -229,14 +298,14 @@ func Test_SingleNodeConcurrentRequests(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to count records: %s", err.Error())
 	}
-	if r != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[200]]}]}` {
+	if r != `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[200]]}]}` {
 		t.Fatalf("test received wrong result got %s", r)
 	}
 }
 
 func Test_SingleNodeConcurrentRequestsCompressed(t *testing.T) {
 	var err error
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	node.Store.SetRequestCompression(0, 0) // Ensure compression
 	defer node.Deprovision()
 
@@ -262,13 +331,13 @@ func Test_SingleNodeConcurrentRequestsCompressed(t *testing.T) {
 	if err != nil {
 		t.Fatalf("failed to count records: %s", err.Error())
 	}
-	if r != `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[200]]}]}` {
+	if r != `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[200]]}]}` {
 		t.Fatalf("test received wrong result got %s", r)
 	}
 }
 
 func Test_SingleNodeParameterized(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	tests := []struct {
@@ -310,8 +379,41 @@ func Test_SingleNodeParameterized(t *testing.T) {
 	}
 }
 
+func Test_SingleNodeRequestParameterized(t *testing.T) {
+	node := mustNewLeaderNode("leader1")
+	defer node.Deprovision()
+
+	tests := []struct {
+		stmt     []interface{}
+		expected string
+	}{
+		{
+			stmt:     []interface{}{"CREATE TABLE foo (id integer not null primary key, name text, age integer)"},
+			expected: `{"results":[{}]}`,
+		},
+		{
+			stmt:     []interface{}{"INSERT INTO foo(name, age) VALUES(?, ?)", "fiona", 20},
+			expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
+		},
+		{
+			stmt:     []interface{}{"SELECT * FROM foo WHERE NAME=?", "fiona"},
+			expected: `{"results":[{"columns":["id","name","age"],"types":["integer","text","integer"],"values":[[1,"fiona",20]]}]}`,
+		},
+	}
+
+	for i, tt := range tests {
+		r, err := node.RequestMultiParameterized(tt.stmt)
+		if err != nil {
+			t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
+		}
+		if r != tt.expected {
+			t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
+		}
+	}
+}
+
 func Test_SingleNodeParameterizedNull(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	tests := []struct {
@@ -354,7 +456,7 @@ func Test_SingleNodeParameterizedNull(t *testing.T) {
 }
 
 func Test_SingleNodeParameterizedNamed(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	tests := []struct {
@@ -399,7 +501,7 @@ func Test_SingleNodeParameterizedNamed(t *testing.T) {
 // Test_SingleNodeParameterizedNamedConstraints tests that named parameters can be used with constraints
 // See https://github.com/rqlite/rqlite/issues/1177
 func Test_SingleNodeParameterizedNamedConstraints(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	_, err := node.ExecuteParameterized([]interface{}{"CREATE TABLE [TestTable] ([Id] int primary key, [Col1] int not null, [Col2] varchar(500), [Col3] int not null, [Col4] varchar(500))"})
@@ -419,7 +521,7 @@ func Test_SingleNodeParameterizedNamedConstraints(t *testing.T) {
 }
 
 func Test_SingleNodeRewriteRandom(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	_, err := node.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`)
@@ -439,7 +541,7 @@ func Test_SingleNodeRewriteRandom(t *testing.T) {
 }
 
 func Test_SingleNodeQueued(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	_, err := node.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`)
@@ -470,7 +572,7 @@ LOOP:
 			if err != nil {
 				t.Fatalf(`query failed: %s`, err.Error())
 			}
-			if r == `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[3]]}]}` {
+			if r == `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[3]]}]}` {
 				break LOOP
 			}
 		case <-timer.C:
@@ -492,14 +594,14 @@ LOOP:
 	if err != nil {
 		t.Fatalf(`query failed: %s`, err.Error())
 	}
-	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[6]]}]}`; got != exp {
+	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[6]]}]}`; got != exp {
 		t.Fatalf("incorrect results, exp: %s, got: %s", exp, got)
 	}
 }
 
 // Test_SingleNodeQueuedBadStmt tests that a single bad SQL statement has the right outcome.
 func Test_SingleNodeQueuedBadStmt(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 	node.Service.DefaultQueueTx = false
 
@@ -535,7 +637,7 @@ LOOP1:
 			if err != nil {
 				t.Fatalf(`query failed: %s`, err.Error())
 			}
-			if r == `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[2]]}]}` {
+			if r == `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[2]]}]}` {
 				break LOOP1
 			}
 		case <-timer.C:
@@ -564,7 +666,7 @@ LOOP2:
 			if err != nil {
 				t.Fatalf(`query failed: %s`, err.Error())
 			}
-			if r == `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[2]]}]}` {
+			if r == `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[2]]}]}` {
 				break LOOP2
 			}
 		case <-timer.C:
@@ -575,7 +677,7 @@ LOOP2:
 }
 
 func Test_SingleNodeQueuedEmptyNil(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	_, err := node.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`)
@@ -609,7 +711,7 @@ func Test_SingleNodeQueuedEmptyNil(t *testing.T) {
 	if err != nil {
 		t.Fatalf(`query failed: %s`, err.Error())
 	}
-	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[3]]}]}`; got != exp {
+	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[3]]}]}`; got != exp {
 		t.Fatalf("incorrect results, exp: %s, got: %s", exp, got)
 	}
 
@@ -634,7 +736,7 @@ func Test_SingleNodeQueuedEmptyNil(t *testing.T) {
 	if err != nil {
 		t.Fatalf(`query failed: %s`, err.Error())
 	}
-	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[6]]}]}`; got != exp {
+	if got, exp := r, `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[6]]}]}`; got != exp {
 		t.Fatalf("incorrect results, exp: %s, got: %s", exp, got)
 	}
 }
@@ -642,7 +744,7 @@ func Test_SingleNodeQueuedEmptyNil(t *testing.T) {
 // Test_SingleNodeSQLInjection demonstrates that using the non-parameterized API is vulnerable to
 // SQL injection attacks.
 func Test_SingleNodeSQLInjection(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	tests := []struct {
@@ -697,7 +799,7 @@ func Test_SingleNodeSQLInjection(t *testing.T) {
 // Test_SingleNodeNoSQLInjection demonstrates that using the parameterized API protects
 // against SQL injection attacks.
 func Test_SingleNodeNoSQLInjection(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	tests := []struct {
@@ -744,12 +846,18 @@ func Test_SingleNodeNoSQLInjection(t *testing.T) {
 	}
 }
 
-// Test_SingleNodeUpgrades upgrade from a data created by earlier releases.
-func Test_SingleNodeUpgrades(t *testing.T) {
-	upgradeTo := func(dir string) {
+// Test_SingleNodeUpgrades_NoSnapshots upgrade from a data created by earlier releases, but which
+// do not have snapshots.
+func Test_SingleNodeUpgrades_NoSnapshots(t *testing.T) {
+	versions := []string{
+		"v7.0.0-data",
+		"v7.9.2-data",
+	}
+
+	upgradeFrom := func(dir string) {
 		// Deprovision of a node deletes the node's dir, so make a copy first.
 		srcdir := filepath.Join("testdata", dir)
-		destdir := mustTempDir()
+		destdir := mustTempDir("")
 		if err := os.Remove(destdir); err != nil {
 			t.Fatalf("failed to remove dest dir: %s", err)
 		}
@@ -759,8 +867,9 @@ func Test_SingleNodeUpgrades(t *testing.T) {
 
 		mux, ln := mustNewOpenMux("")
 		defer ln.Close()
-
-		node := mustNodeEncrypted(destdir, true, false, mux, "node1")
+		raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+		clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+		node := mustNodeEncrypted("node1", destdir, true, false, mux, raftDialer, clstrDialer)
 		defer node.Deprovision()
 		if _, err := node.WaitForLeader(); err != nil {
 			t.Fatalf("node never became leader with %s data:", dir)
@@ -783,7 +892,7 @@ func Test_SingleNodeUpgrades(t *testing.T) {
 				if err != nil {
 					t.Fatalf("query failed with %s data: %s", dir, err)
 				}
-				expected := `{"results":[{"columns":["COUNT(*)"],"types":[""],"values":[[20]]}]}`
+				expected := `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[20]]}]}`
 				if r == expected {
 					close(testSuccess)
 				}
@@ -791,13 +900,154 @@ func Test_SingleNodeUpgrades(t *testing.T) {
 		}
 	}
 
-	upgradeTo("v6.0.0-data")
-	upgradeTo("v7.0.0-data")
-	upgradeTo("v7.9.2-data")
+	for _, version := range versions {
+		t.Run(version, func(t *testing.T) {
+			upgradeFrom(version)
+		})
+	}
+}
+
+// Test_SingleNodeUpgrades_Snapshots upgrade from a data created by earlier releases, but which
+// do have snapshots.
+func Test_SingleNodeUpgrades_Snapshots(t *testing.T) {
+	versions := []string{
+		"v7.20.3-data-with-snapshots",
+	}
+
+	upgradeFrom := func(dir string) {
+		// Deprovision of a node deletes the node's dir, so make a copy first.
+		srcdir := filepath.Join("testdata", dir)
+		destdir := mustTempDir("")
+		if err := os.Remove(destdir); err != nil {
+			t.Fatalf("failed to remove dest dir: %s", err)
+		}
+		if err := copyDir(srcdir, destdir); err != nil {
+			t.Fatalf("failed to copy node test directory: %s", err)
+		}
+
+		mux, ln := mustNewOpenMux("")
+		defer ln.Close()
+		raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+		clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+		node := mustNodeEncrypted("node1", destdir, true, false, mux, raftDialer, clstrDialer)
+		defer node.Deprovision()
+		if _, err := node.WaitForLeader(); err != nil {
+			t.Fatalf("node never became leader with %s data:", dir)
+		}
+
+		timer := time.NewTimer(5 * time.Second)
+		defer timer.Stop()
+		ticker := time.NewTicker(100 * time.Millisecond)
+		defer ticker.Stop()
+		testSuccess := make(chan struct{})
+
+		for {
+			select {
+			case <-testSuccess:
+				return
+			case <-timer.C:
+				t.Fatalf(`timeout waiting for correct results with %s data`, dir)
+			case <-ticker.C:
+				r, err := node.QueryNoneConsistency(`SELECT COUNT(*) FROM foo`)
+				if err != nil {
+					t.Fatalf("query failed with %s data: %s", dir, err)
+				}
+				expected := `{"results":[{"columns":["COUNT(*)"],"types":["integer"],"values":[[20]]}]}`
+				if r == expected {
+					close(testSuccess)
+				}
+			}
+		}
+	}
+
+	for _, version := range versions {
+		t.Run(version, func(t *testing.T) {
+			upgradeFrom(version)
+		})
+	}
+}
+
+func Test_SingleNodeBackup(t *testing.T) {
+	node := mustNewLeaderNode("leader1")
+	defer node.Deprovision()
+
+	// create a table and insert a row
+	_, err := node.Execute(`CREATE TABLE foo (id integer not null primary key, name text)`)
+	if err != nil {
+		t.Fatalf(`CREATE TABLE failed: %s`, err.Error())
+	}
+	_, err = node.Execute(`INSERT INTO foo(name) VALUES("fiona")`)
+	if err != nil {
+		t.Fatalf(`INSERT failed: %s`, err.Error())
+	}
+
+	backup := mustTempFile()
+	defer os.Remove(backup)
+	if err := node.Backup(backup, false); err != nil {
+		t.Fatalf(`backup failed: %s`, err.Error())
+	}
+
+	// Create new node, and restore from backup.
+	newNode := mustNewLeaderNode("leader2")
+	defer newNode.Deprovision()
+	_, err = newNode.Boot(backup)
+	if err != nil {
+		t.Fatalf(`boot failed: %s`, err.Error())
+	}
+	r, err := newNode.Query(`SELECT * FROM foo`)
+	if err != nil {
+		t.Fatalf(`query failed: %s`, err.Error())
+	}
+	if r != `{"results":[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]}` {
+		t.Fatalf("test received wrong result got %s", r)
+	}
+
+	// Get a compressed backup, test it again.
+	if err := node.Backup(backup, true); err != nil {
+		t.Fatalf(`backup failed: %s`, err.Error())
+	}
+
+	// decompress backup and check it.
+	decompressedBackup := mustTempFile()
+	defer os.Remove(decompressedBackup)
+
+	f, err := os.Open(backup)
+	if err != nil {
+		t.Fatalf(`failed to open backup: %s`, err.Error())
+	}
+	defer f.Close()
+	gzr, err := gzip.NewReader(f)
+	if err != nil {
+		t.Fatalf(`failed to create gzip reader: %s`, err.Error())
+	}
+	defer gzr.Close()
+	w, err := os.Create(decompressedBackup)
+	if err != nil {
+		t.Fatalf(`failed to create decompressed backup: %s`, err.Error())
+	}
+	defer w.Close()
+	if _, err := io.Copy(w, gzr); err != nil {
+		t.Fatalf(`failed to decompress backup: %s`, err.Error())
+	}
+
+	// Create new node, and restore from backup.
+	newNode2 := mustNewLeaderNode("leader3")
+	defer newNode2.Deprovision()
+	_, err = newNode2.Boot(decompressedBackup)
+	if err != nil {
+		t.Fatalf(`boot failed: %s`, err.Error())
+	}
+	r, err = newNode2.Query(`SELECT * FROM foo`)
+	if err != nil {
+		t.Fatalf(`query failed: %s`, err.Error())
+	}
+	if r != `{"results":[{"columns":["id","name"],"types":["integer","text"],"values":[[1,"fiona"]]}]}` {
+		t.Fatalf("test received wrong result got %s", r)
+	}
 }
 
 func Test_SingleNodeNodes(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	// Access endpoints to ensure the code is covered.
@@ -809,10 +1059,7 @@ func Test_SingleNodeNodes(t *testing.T) {
 	if len(nodes) != 1 {
 		t.Fatalf("wrong number of nodes in response")
 	}
-	n, ok := nodes[node.ID]
-	if !ok {
-		t.Fatalf("node not found by ID in response")
-	}
+	n := nodes[0]
 	if n.Addr != node.RaftAddr {
 		t.Fatalf("node has wrong Raft address")
 	}
@@ -828,7 +1075,7 @@ func Test_SingleNodeNodes(t *testing.T) {
 }
 
 func Test_SingleNodeCoverage(t *testing.T) {
-	node := mustNewLeaderNode()
+	node := mustNewLeaderNode("leader1")
 	defer node.Deprovision()
 
 	// Access endpoints to ensure the code is covered.
@@ -854,135 +1101,117 @@ func Test_SingleNodeCoverage(t *testing.T) {
 
 // Test_SingleNodeReopen tests that a node can be re-opened OK.
 func Test_SingleNodeReopen(t *testing.T) {
-	onDisk := false
+	dir := mustTempDir("node1")
+	mux, ln := mustNewOpenMux("")
+	defer ln.Close()
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+	node := mustNodeEncrypted("node1", dir, true, false, mux, raftDialer, clstrDialer)
 
-	for {
-		t.Logf("running test %s, on-disk=%v", t.Name(), onDisk)
-
-		dir := mustTempDir()
-		mux, ln := mustNewOpenMux("")
-		defer ln.Close()
-		node := mustNodeEncrypted(dir, true, false, mux, "")
-
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
-		}
-
-		if err := node.Close(true); err != nil {
-			t.Fatalf("failed to close node")
-		}
-
-		if err := node.Store.Open(); err != nil {
-			t.Fatalf("failed to re-open store: %s", err)
-		}
-		if err := node.Store.Bootstrap(store.NewServer(node.Store.ID(), node.Store.Addr(), true)); err != nil {
-			t.Fatalf("failed to bootstrap single-node store: %s", err.Error())
-		}
-		if err := node.Service.Start(); err != nil {
-			t.Fatalf("failed to restart service: %s", err)
-		}
-
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
-		}
-
-		node.Deprovision()
-		// Switch to other mode for another test.
-		onDisk = !onDisk
-		if onDisk == false {
-			break
-		}
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
 	}
+
+	if err := node.Close(true); err != nil {
+		t.Fatalf("failed to close node")
+	}
+
+	if err := node.Store.Open(); err != nil {
+		t.Fatalf("failed to re-open store: %s", err)
+	}
+	if err := node.Store.Bootstrap(store.NewServer(node.Store.ID(), node.Store.Addr(), true)); err != nil {
+		t.Fatalf("failed to bootstrap single-node store: %s", err.Error())
+	}
+	if err := node.Service.Start(); err != nil {
+		t.Fatalf("failed to restart service: %s", err)
+	}
+
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
+
+	node.Deprovision()
 }
 
 // Test_SingleNodeReopen tests that a node can be re-opened OK, with
 // a non-database command in the log.
 func Test_SingleNodeNoopReopen(t *testing.T) {
-	onDisk := false
+	dir := mustTempDir("node1")
+	mux, ln := mustNewOpenMux("")
+	defer ln.Close()
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+	node := mustNodeEncrypted("node1", dir, true, false, mux, raftDialer, clstrDialer)
 
-	for {
-		t.Logf("running test %s, on-disk=%v", t.Name(), onDisk)
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
 
-		dir := mustTempDir()
-		mux, ln := mustNewOpenMux("")
-		defer ln.Close()
-		node := mustNodeEncryptedOnDisk(dir, true, false, mux, "", false)
+	if err := node.Noop("#1"); err != nil {
+		t.Fatalf("failed to write noop command: %s", err)
+	}
 
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
+	if err := node.Close(true); err != nil {
+		t.Fatalf("failed to close node")
+	}
+
+	if err := node.Store.Open(); err != nil {
+		t.Fatalf("failed to re-open store: %s", err)
+	}
+	if err := node.Store.Bootstrap(store.NewServer(node.Store.ID(), node.Store.Addr(), true)); err != nil {
+		t.Fatalf("failed to bootstrap single-node store: %s", err.Error())
+	}
+	if err := node.Service.Start(); err != nil {
+		t.Fatalf("failed to restart service: %s", err)
+	}
+	// This testing tells service to restart with localhost:0
+	// again, so explicitly set the API address again.
+	node.APIAddr = node.Service.Addr().String()
+
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
+
+	// Ensure node is fully functional after restart.
+	tests := []struct {
+		stmt     []interface{}
+		expected string
+		execute  bool
+	}{
+		{
+			stmt:     []interface{}{"CREATE TABLE foo (id integer not null primary key, name text, age integer)"},
+			expected: `{"results":[{}]}`,
+			execute:  true,
+		},
+		{
+			stmt:     []interface{}{"INSERT INTO foo(name, age) VALUES(?, ?)", "fiona", 20},
+			expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
+			execute:  true,
+		},
+		{
+			stmt:     []interface{}{"SELECT * FROM foo WHERE NAME=?", "fiona"},
+			expected: `{"results":[{"columns":["id","name","age"],"types":["integer","text","integer"],"values":[[1,"fiona",20]]}]}`,
+			execute:  false,
+		},
+	}
+
+	for i, tt := range tests {
+		var r string
+		var err error
+		if tt.execute {
+			r, err = node.ExecuteParameterized(tt.stmt)
+		} else {
+			r, err = node.QueryParameterized(tt.stmt)
 		}
-
-		if err := node.Noop("#1"); err != nil {
-			t.Fatalf("failed to write noop command: %s", err)
+		if err != nil {
+			t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
 		}
-
-		if err := node.Close(true); err != nil {
-			t.Fatalf("failed to close node")
-		}
-
-		if err := node.Store.Open(); err != nil {
-			t.Fatalf("failed to re-open store: %s", err)
-		}
-		if err := node.Store.Bootstrap(store.NewServer(node.Store.ID(), node.Store.Addr(), true)); err != nil {
-			t.Fatalf("failed to bootstrap single-node store: %s", err.Error())
-		}
-		if err := node.Service.Start(); err != nil {
-			t.Fatalf("failed to restart service: %s", err)
-		}
-		// This testing tells service to restart with localhost:0
-		// again, so explicitly set the API address again.
-		node.APIAddr = node.Service.Addr().String()
-
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
-		}
-
-		// Ensure node is fully functional after restart.
-		tests := []struct {
-			stmt     []interface{}
-			expected string
-			execute  bool
-		}{
-			{
-				stmt:     []interface{}{"CREATE TABLE foo (id integer not null primary key, name text, age integer)"},
-				expected: `{"results":[{}]}`,
-				execute:  true,
-			},
-			{
-				stmt:     []interface{}{"INSERT INTO foo(name, age) VALUES(?, ?)", "fiona", 20},
-				expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
-				execute:  true,
-			},
-			{
-				stmt:     []interface{}{"SELECT * FROM foo WHERE NAME=?", "fiona"},
-				expected: `{"results":[{"columns":["id","name","age"],"types":["integer","text","integer"],"values":[[1,"fiona",20]]}]}`,
-				execute:  false,
-			},
-		}
-
-		for i, tt := range tests {
-			var r string
-			var err error
-			if tt.execute {
-				r, err = node.ExecuteParameterized(tt.stmt)
-			} else {
-				r, err = node.QueryParameterized(tt.stmt)
-			}
-			if err != nil {
-				t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
-			}
-			if r != tt.expected {
-				t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
-			}
-		}
-
-		node.Deprovision()
-		// Switch to other mode for another test.
-		onDisk = !onDisk
-		if onDisk == false {
-			break
+		if r != tt.expected {
+			t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
 		}
 	}
+
+	node.Deprovision()
 }
 
 // Test_SingleNodeReopen tests that a node can be re-opened OK, with
@@ -990,95 +1219,87 @@ func Test_SingleNodeNoopReopen(t *testing.T) {
 // This tests that the code can handle a snapshot that doesn't
 // contain database data. This shouldn't happen in real systems
 func Test_SingleNodeNoopSnapReopen(t *testing.T) {
-	onDisk := false
+	dir := mustTempDir("node1")
+	mux, ln := mustNewOpenMux("")
+	defer ln.Close()
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+	node := mustNodeEncrypted("node1", dir, true, false, mux, raftDialer, clstrDialer)
 
-	for {
-		t.Logf("running test %s, on-disk=%v", t.Name(), onDisk)
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
 
-		dir := mustTempDir()
-		mux, ln := mustNewOpenMux("")
-		defer ln.Close()
-		node := mustNodeEncryptedOnDisk(dir, true, false, mux, "", onDisk)
-
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
-		}
-
-		for i := 0; i < 150; i++ {
-			if err := node.Noop(fmt.Sprintf("%d", i)); err != nil {
-				t.Fatalf("failed to write noop command: %s", err)
-			}
-		}
-
-		// Wait for a snapshot to happen.
-		time.Sleep(5 * time.Second)
-
-		if err := node.Close(true); err != nil {
-			t.Fatalf("failed to close node")
-		}
-
-		if err := node.Store.Open(); err != nil {
-			t.Fatalf("failed to re-open store: %s", err)
-		}
-		if err := node.Store.Bootstrap(store.NewServer(node.Store.ID(), node.Store.Addr(), true)); err != nil {
-			t.Fatalf("failed to bootstrap single-node store: %s", err.Error())
-		}
-		if err := node.Service.Start(); err != nil {
-			t.Fatalf("failed to restart service: %s", err)
-		}
-		// This testing tells service to restart with localhost:0
-		// again, so explicitly set the API address again.
-		node.APIAddr = node.Service.Addr().String()
-
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
-		}
-
-		// Ensure node is fully functional after restart.
-		tests := []struct {
-			stmt     []interface{}
-			expected string
-			execute  bool
-		}{
-			{
-				stmt:     []interface{}{"CREATE TABLE foo (id integer not null primary key, name text, age integer)"},
-				expected: `{"results":[{}]}`,
-				execute:  true,
-			},
-			{
-				stmt:     []interface{}{"INSERT INTO foo(name, age) VALUES(?, ?)", "fiona", 20},
-				expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
-				execute:  true,
-			},
-			{
-				stmt:     []interface{}{"SELECT * FROM foo WHERE NAME=?", "fiona"},
-				expected: `{"results":[{"columns":["id","name","age"],"types":["integer","text","integer"],"values":[[1,"fiona",20]]}]}`,
-				execute:  false,
-			},
-		}
-
-		for i, tt := range tests {
-			var r string
-			var err error
-			if tt.execute {
-				r, err = node.ExecuteParameterized(tt.stmt)
-			} else {
-				r, err = node.QueryParameterized(tt.stmt)
-			}
-			if err != nil {
-				t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
-			}
-			if r != tt.expected {
-				t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
-			}
-		}
-
-		node.Deprovision()
-		onDisk = !onDisk
-		if onDisk == false {
-			break
+	for i := 0; i < 150; i++ {
+		if err := node.Noop(fmt.Sprintf("%d", i)); err != nil {
+			t.Fatalf("failed to write noop command: %s", err)
 		}
 	}
+
+	// Wait for a snapshot to happen.
+	time.Sleep(5 * time.Second)
+
+	if err := node.Close(true); err != nil {
+		t.Fatalf("failed to close node")
+	}
+
+	if err := node.Store.Open(); err != nil {
+		t.Fatalf("failed to re-open store: %s", err)
+	}
+	if err := node.Store.Bootstrap(store.NewServer(node.Store.ID(), node.Store.Addr(), true)); err != nil {
+		t.Fatalf("failed to bootstrap single-node store: %s", err.Error())
+	}
+	if err := node.Service.Start(); err != nil {
+		t.Fatalf("failed to restart service: %s", err)
+	}
+	// This testing tells service to restart with localhost:0
+	// again, so explicitly set the API address again.
+	node.APIAddr = node.Service.Addr().String()
+
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
+
+	// Ensure node is fully functional after restart.
+	tests := []struct {
+		stmt     []interface{}
+		expected string
+		execute  bool
+	}{
+		{
+			stmt:     []interface{}{"CREATE TABLE foo (id integer not null primary key, name text, age integer)"},
+			expected: `{"results":[{}]}`,
+			execute:  true,
+		},
+		{
+			stmt:     []interface{}{"INSERT INTO foo(name, age) VALUES(?, ?)", "fiona", 20},
+			expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
+			execute:  true,
+		},
+		{
+			stmt:     []interface{}{"SELECT * FROM foo WHERE NAME=?", "fiona"},
+			expected: `{"results":[{"columns":["id","name","age"],"types":["integer","text","integer"],"values":[[1,"fiona",20]]}]}`,
+			execute:  false,
+		},
+	}
+
+	for i, tt := range tests {
+		var r string
+		var err error
+		if tt.execute {
+			r, err = node.ExecuteParameterized(tt.stmt)
+		} else {
+			r, err = node.QueryParameterized(tt.stmt)
+		}
+		if err != nil {
+			t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
+		}
+		if r != tt.expected {
+			t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
+		}
+	}
+
+	node.Deprovision()
 }
 
 // Test_SingleNodeNoopSnapLogsReopen tests that a node can be re-opened OK,
@@ -1086,99 +1307,180 @@ func Test_SingleNodeNoopSnapReopen(t *testing.T) {
 // This tests that the code can handle a snapshot that doesn't
 // contain database data. This shouldn't happen in real systems
 func Test_SingleNodeNoopSnapLogsReopen(t *testing.T) {
-	onDisk := false
-	var raftAddr string
+	dir := mustTempDir("node1")
+	mux, ln := mustNewOpenMux("")
+	defer ln.Close()
+	raftDialer := tcp.NewDialer(cluster.MuxRaftHeader, nil)
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+	node := mustNodeEncrypted("node1", dir, true, false, mux, raftDialer, clstrDialer)
 
-	for {
-		t.Logf("running test %s, on-disk=%v", t.Name(), onDisk)
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
 
-		dir := mustTempDir()
-		mux, ln := mustNewOpenMux("")
-		defer ln.Close()
-		node := mustNodeEncryptedOnDisk(dir, true, false, mux, "", onDisk)
-		raftAddr = node.RaftAddr
-		t.Logf("node listening for Raft on %s", raftAddr)
-
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
+	for i := 0; i < 150; i++ {
+		if err := node.Noop(fmt.Sprintf("%d", i)); err != nil {
+			t.Fatalf("failed to write noop command: %s", err)
 		}
+	}
 
-		for i := 0; i < 150; i++ {
-			if err := node.Noop(fmt.Sprintf("%d", i)); err != nil {
-				t.Fatalf("failed to write noop command: %s", err)
-			}
+	// Wait for a snapshot to happen, and then write some more commands.
+	time.Sleep(5 * time.Second)
+	for i := 0; i < 5; i++ {
+		if err := node.Noop(fmt.Sprintf("%d", i)); err != nil {
+			t.Fatalf("failed to write noop command: %s", err)
 		}
+	}
 
-		// Wait for a snapshot to happen, and then write some more commands.
-		time.Sleep(5 * time.Second)
-		for i := 0; i < 5; i++ {
-			if err := node.Noop(fmt.Sprintf("%d", i)); err != nil {
-				t.Fatalf("failed to write noop command: %s", err)
-			}
-		}
+	if err := node.Close(true); err != nil {
+		t.Fatalf("failed to close node")
+	}
 
-		if err := node.Close(true); err != nil {
-			t.Fatalf("failed to close node")
-		}
+	if err := node.Store.Open(); err != nil {
+		t.Fatalf("failed to re-open store: %s", err)
+	}
+	if err := node.Service.Start(); err != nil {
+		t.Fatalf("failed to restart service: %s", err)
+	}
+	// This testing tells service to restart with localhost:0
+	// again, so explicitly set the API address again.
+	node.APIAddr = node.Service.Addr().String()
 
-		if err := node.Store.Open(); err != nil {
-			t.Fatalf("failed to re-open store: %s", err)
-		}
-		if err := node.Service.Start(); err != nil {
-			t.Fatalf("failed to restart service: %s", err)
-		}
-		// This testing tells service to restart with localhost:0
-		// again, so explicitly set the API address again.
-		node.APIAddr = node.Service.Addr().String()
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
 
-		if _, err := node.WaitForLeader(); err != nil {
-			t.Fatalf("node never became leader")
-		}
+	// Ensure node is fully functional after restart.
+	tests := []struct {
+		stmt     []interface{}
+		expected string
+		execute  bool
+	}{
+		{
+			stmt:     []interface{}{"CREATE TABLE foo (id integer not null primary key, name text, age integer)"},
+			expected: `{"results":[{}]}`,
+			execute:  true,
+		},
+		{
+			stmt:     []interface{}{"INSERT INTO foo(name, age) VALUES(?, ?)", "fiona", 20},
+			expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
+			execute:  true,
+		},
+		{
+			stmt:     []interface{}{"SELECT * FROM foo WHERE NAME=?", "fiona"},
+			expected: `{"results":[{"columns":["id","name","age"],"types":["integer","text","integer"],"values":[[1,"fiona",20]]}]}`,
+			execute:  false,
+		},
+	}
 
-		// Ensure node is fully functional after restart.
-		tests := []struct {
-			stmt     []interface{}
-			expected string
-			execute  bool
-		}{
-			{
-				stmt:     []interface{}{"CREATE TABLE foo (id integer not null primary key, name text, age integer)"},
-				expected: `{"results":[{}]}`,
-				execute:  true,
-			},
-			{
-				stmt:     []interface{}{"INSERT INTO foo(name, age) VALUES(?, ?)", "fiona", 20},
-				expected: `{"results":[{"last_insert_id":1,"rows_affected":1}]}`,
-				execute:  true,
-			},
-			{
-				stmt:     []interface{}{"SELECT * FROM foo WHERE NAME=?", "fiona"},
-				expected: `{"results":[{"columns":["id","name","age"],"types":["integer","text","integer"],"values":[[1,"fiona",20]]}]}`,
-				execute:  false,
-			},
+	for i, tt := range tests {
+		var r string
+		var err error
+		if tt.execute {
+			r, err = node.ExecuteParameterized(tt.stmt)
+		} else {
+			r, err = node.QueryParameterized(tt.stmt)
 		}
+		if err != nil {
+			t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
+		}
+		if r != tt.expected {
+			t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
+		}
+	}
 
-		for i, tt := range tests {
-			var r string
-			var err error
-			if tt.execute {
-				r, err = node.ExecuteParameterized(tt.stmt)
-			} else {
-				r, err = node.QueryParameterized(tt.stmt)
-			}
-			if err != nil {
-				t.Fatalf(`test %d failed "%s": %s`, i, tt.stmt, err.Error())
-			}
-			if r != tt.expected {
-				t.Fatalf(`test %d received wrong result "%s" got: %s exp: %s`, i, tt.stmt, r, tt.expected)
-			}
-		}
+	node.Deprovision()
+}
 
-		node.Deprovision()
-		// Switch to other mode for another test.
-		onDisk = !onDisk
-		if onDisk == false {
-			break
-		}
+func Test_SingleNodeAutoRestore(t *testing.T) {
+	dir := mustTempDir("node1")
+	node := &Node{
+		Dir:       dir,
+		PeersPath: filepath.Join(dir, "raft/peers.json"),
+	}
+	defer node.Deprovision()
+
+	mux, _ := mustNewOpenMux("")
+	go mux.Serve()
+
+	raftLn := mux.Listen(cluster.MuxRaftHeader)
+	raftTn := tcp.NewLayer(raftLn, tcp.NewDialer(cluster.MuxRaftHeader, nil))
+	node.Store = store.New(raftTn, &store.Config{
+		DBConf: store.NewDBConfig(),
+		Dir:    node.Dir,
+		ID:     "node1",
+	})
+
+	restoreFile := mustTempFile()
+	copyFile(filepath.Join("testdata", "auto-restore.sqlite"), restoreFile)
+	if err := node.Store.SetRestorePath(restoreFile); err != nil {
+		t.Fatalf("failed to set restore path: %s", err.Error())
+	}
+
+	if err := node.Store.Open(); err != nil {
+		t.Fatalf("failed to open store: %s", err.Error())
+	}
+	if err := node.Store.Bootstrap(store.NewServer(node.Store.ID(), node.Store.Addr(), true)); err != nil {
+		t.Fatalf("failed to bootstrap store: %s", err.Error())
+	}
+	node.RaftAddr = node.Store.Addr()
+	node.ID = node.Store.ID()
+
+	clstr := cluster.New(mux.Listen(cluster.MuxClusterHeader), node.Store, node.Store, mustNewMockCredentialStore())
+	if err := clstr.Open(); err != nil {
+		t.Fatalf("failed to open Cluster service: %s", err.Error())
+	}
+	node.Cluster = clstr
+
+	clstrDialer := tcp.NewDialer(cluster.MuxClusterHeader, nil)
+	clstrClient := cluster.NewClient(clstrDialer, 30*time.Second)
+	node.Service = httpd.New("localhost:0", node.Store, clstrClient, nil)
+
+	if err := node.Service.Start(); err != nil {
+		t.Fatalf("failed to start HTTP server: %s", err.Error())
+	}
+	node.APIAddr = node.Service.Addr().String()
+
+	// Finally, set API address in Cluster service
+	clstr.SetAPIAddr(node.APIAddr)
+
+	if _, err := node.WaitForLeader(); err != nil {
+		t.Fatalf("node never became leader")
+	}
+	testPoll(t, node.Ready, 100*time.Millisecond, 5*time.Second)
+
+	r, err := node.Query("SELECT * FROM foo WHERE id=2")
+	if err != nil {
+		t.Fatalf("failed to execute query: %s", err.Error())
+	}
+	if r != `{"results":[{"columns":["id","name"],"types":["integer","text"],"values":[[2,"fiona"]]}]}` {
+		t.Fatalf("test received wrong result got %s", r)
+	}
+}
+
+func Test_SingleNodeBoot_OK(t *testing.T) {
+	node := mustNewLeaderNode("leader1")
+	defer node.Deprovision()
+
+	_, err := node.Boot(filepath.Join("testdata", "auto-restore.sqlite"))
+	if err != nil {
+		t.Fatalf("failed to load data: %s", err.Error())
+	}
+
+	r, err := node.Query("SELECT * FROM foo WHERE id=2")
+	if err != nil {
+		t.Fatalf("failed to execute query: %s", err.Error())
+	}
+	if r != `{"results":[{"columns":["id","name"],"types":["integer","text"],"values":[[2,"fiona"]]}]}` {
+		t.Fatalf("test received wrong result got %s", r)
+	}
+}
+
+func Test_SingleNodeBoot_FailNotLeader(t *testing.T) {
+	node := mustNewNode("node1", false)
+	defer node.Deprovision()
+	_, err := node.Boot(filepath.Join("testdata", "auto-restore.sqlite"))
+	if err == nil {
+		t.Fatalf("expected error loading data")
 	}
 }
